% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboClass.R
\name{aggregate_samples_by_col}
\alias{aggregate_samples_by_col}
\title{Aggregate results into a single matrix}
\usage{
aggregate_samples_by_col(sample_list, col_to_keep, position_to_rownames = F)
}
\arguments{
\item{sample_list}{}

\item{col_to_keep}{}

\item{position_to_rownames}{}
}
\value{

}
\description{
Aggregate results into a single matrix
}
