% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_check.R
\name{batch_report}
\alias{batch_report}
\title{Generate html report about batch effect in the working directory}
\usage{
batch_report(ribo, library_col)
}
\arguments{
\item{ribo}{a riboclass}

\item{library_col}{library column in the metadata}
}
\description{
Generate html report about batch effect in the working directory
}
