% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_heatmap_annotation}
\alias{build_heatmap_annotation}
\title{Build a consistent ComplexHeatmap sample annotation layer.}
\usage{
build_heatmap_annotation(
  metadata = NULL,
  color_col = NULL,
  sample_colors = NULL,
  na_col = "red"
)
}
\arguments{
\item{metadata}{Metadata dataframe.}

\item{color_col}{Metadata columns to annotate, or NULL.}

\item{sample_colors}{Optional custom colors.}

\item{na_col}{Color for missing annotation values.}
}
\description{
Build a consistent ComplexHeatmap sample annotation layer.
}
\keyword{internal}
