% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculate score for a whole sample cohort}
\usage{
calculate_score(
  ribo = NULL,
  flanking = 6,
  data_rna_col = 1,
  data_counts_col = 3,
  median_or_mean = "median",
  use_multithreads = F
)
}
\arguments{
\item{ribo}{a riboclass object}

\item{flanking}{the window size around the position (the latter is excluded)}

\item{data_rna_col}{Name or position of the column containing the RNA}

\item{data_counts_col}{Name or position of the column containing count values}
}
\value{

}
\description{
Calculate score for a whole sample cohort
}
