% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{calculate_score_by_RNA}
\alias{calculate_score_by_RNA}
\title{Calculate both C-score and Z-score for one RNA of one sample.}
\usage{
calculate_score_by_RNA(
  ds = NULL,
  flanking = 6,
  data.counts.col = 4,
  data.position.col = 1
)
}
\arguments{
\item{ds}{}

\item{flanking}{}

\item{data.counts.col}{}

\item{data.position.col}{}
}
\value{

}
\description{
Calculate both C-score and Z-score for one RNA of one sample.
}
