% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{calculate_score_by_sample}
\alias{calculate_score_by_sample}
\title{Title}
\usage{
calculate_score_by_sample(
  sample_df = NULL,
  data_rna_col = 1,
  flanking = 6,
  data_counts_col = 3
)
}
\arguments{
\item{sample_df}{}

\item{data_rna_col}{}

\item{flanking}{}

\item{data_counts_col}{}
}
\value{

}
\description{
Title
}
