% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_in_set}
\alias{check_in_set}
\title{Check if an argument is in a allowed set of values}
\usage{
check_in_set(x, set, name)
}
\arguments{
\item{x}{The argument to check.}

\item{set}{The set of allowed values.}

\item{name}{Name of the argument (for error message).}
}
\description{
Check if an argument is in a allowed set of values
}
\keyword{internal}
