% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_is_se}
\alias{check_is_se}
\title{Check if an object is a SummarizedExperiment}
\usage{
check_is_se(object)
}
\arguments{
\item{object}{The object to check.}
}
\description{
Check if an object is a SummarizedExperiment
}
\keyword{internal}
