% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_named_colors}
\alias{check_named_colors}
\title{Validate a named color mapping against observed metadata values.}
\usage{
check_named_colors(colors, values, arg_name = "sample_colors")
}
\arguments{
\item{colors}{Named character vector of colors.}

\item{values}{Metadata values that will be mapped to colors.}

\item{arg_name}{Argument name used in error messages.}
}
\description{
Validate a named color mapping against observed metadata values.
}
\keyword{internal}
