% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{check_sample_positions}
\alias{check_sample_positions}
\title{Check if two samples share the same positions}
\usage{
check_sample_positions(sample_1, sample_2)
}
\arguments{
\item{sample_1}{First sample to check against sample_2}

\item{sample_2}{Second sample to check against sample_1}
}
\value{
A boolean indicating if the two samples are identical.
}
\description{
Check if two samples share the same positions
}
