% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_type}
\alias{check_type}
\title{Check the type of an argument}
\usage{
check_type(x, type, name, length = NULL)
}
\arguments{
\item{x}{The argument to check.}

\item{type}{Expected type ("character", "numeric", "logical", "data.frame", "list").}

\item{name}{Name of the argument (for error message).}

\item{length}{Expected length (optional).}
}
\description{
Check the type of an argument
}
\keyword{internal}
