% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{compute.cscore.zscore}
\alias{compute.cscore.zscore}
\title{Computes C- and Z-scores of all samples of a given study}
\usage{
compute.cscore.zscore(all.samples = NULL, annot = NULL, annot.rna.col = 4,
  species = "human", verbose = FALSE, window.size = 6,
  data.position.col = 1, data.counts.col = 4, gc.probas.pos.col = 1,
  gc.probas.GC.col = 7)
}
\arguments{
\item{gc.probas.GC.col}{}

\item{annot:}{a data frame with general project informations. Must have columns specifying the project, rRNA, sample names, paths.}

\item{annot.rna.col:}{a numerical value specifying the column of the rRNA in the annotation <annot>. Default is 4.}

\item{species:}{a character string specifying the species. Default is "human". In this version of the R-packge, only "human" is supported.}
}
\value{
List of data frames. Each data frame is a sample (count table) for which the C- and Z-scores have been computed
}
\description{
Computes C- and Z-scores of all samples of a given study
}
\examples{
compute.cscore.zscore(all.samples, annot)
}
