% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{compute.window.median.mad}
\alias{compute.window.median.mad}
\title{Computation of C- and Z-scores for genomic positions within a predefined window}
\usage{
compute.window.median.mad(ds = NULL, known.meth.vector = NULL,
  suspected.meth.vector = NULL, flanking = 6, data.counts.col = 4,
  data.position.col = 1)
}
\arguments{
\item{data.position.col}{}
}
\value{
data frame with C-scores and Z-scores for each genomic postion
}
\description{
Computation of C- and Z-scores for genomic positions within a predefined window
}
\examples{
compute.window.median.mad(ds, known.meth.vector, suspected.meth.vector)
}
