% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pval.R
\name{compute_pval}
\alias{compute_pval}
\title{Compute statistics (p-value for, Welch, Wilcoxon, Kruskal, ANOVA test) for C-score based on condition column}
\usage{
compute_pval(ribo = ribo, test = NULL, condition_col = NULL)
}
\arguments{
\item{ribo}{a RiboClass object}

\item{test}{statistic test wanted student, anova or kruskal}

\item{condition_col}{column condition in metadata}
}
\value{
a dataframe
}
\description{
Compute statistics (p-value for, Welch, Wilcoxon, Kruskal, ANOVA test) for C-score based on condition column
}
\examples{
data("ribo_toy")
compute_pval(ribo = ribo_toy, test = "student", condition_col = "condition")
}
