% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_window_score_cpp}
\alias{compute_window_score_cpp}
\title{Compute C-score window values (Optimized)}
\usage{
compute_window_score_cpp(count, flanking, method)
}
\arguments{
\item{count}{Vector of counts}

\item{flanking}{Size of the flanking window}

\item{method}{"mean" or "median"}
}
\value{
Vector of flanking values (same length as count)
}
\description{
Compute C-score window values (Optimized)
}
