% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_se.R
\name{create_se}
\alias{create_se}
\title{Create a RiboClass from count files and metadata.}
\usage{
create_se(
  count_path,
  metadata = NULL,
  count_sep = "\\t",
  metadata_sep = ",",
  count_header = FALSE,
  count_value = 3,
  count_rnaid = 1,
  count_pos = 2,
  metadata_key = "filename",
  metadata_id = NULL
)
}
\arguments{
\item{count_path}{(required) path to the data folder containing count files.}

\item{metadata}{Data frame or path to a CSV file containing metadata.}

\item{count_sep}{Delimiter used for the CSV files.}

\item{metadata_sep}{Delimiter used in metadata (if imported from file).}

\item{count_header}{Boolean, specify if count files have a header or not.}

\item{count_value}{Name or index of the column containing count values.}

\item{count_rnaid}{Name or index of the column containing the name of the RNA in count data.}

\item{count_pos}{Name or index of the column containing the site's position in count data.}

\item{metadata_key}{(required when metadata not null) Name or index of the column containing the samples' filename.}

\item{metadata_id}{Name or index of the column containing the sample name.}
}
\value{
A SummarizedExperiment object.
}
\description{
Read ribomethseq count files and their associated metadata and merge them into a SummarizedExperiment.
}
\details{
The resulting \code{SummarizedExperiment} object contains:
\itemize{
\item \code{assays}: A list containing the \code{counts} matrix.
\item \code{rowData}: A \code{DataFrame} containing information for each site:
\itemize{
\item \code{rna}: The name of the RNA.
\item \code{rnapos}: The position on the RNA.
\item \code{site}: The site annotation (if applicable).
}
\item \code{colData}: A \code{DataFrame} containing metadata for each sample.
\item \code{metadata}: A list containing \code{rna_names} table.
}

The path given in \code{count_path} should contain only necessary count files (one per sample).
While the directory structure is not important, make sure each sample has a unique filename.

The \code{metadata} must contain a column (specified by \code{metadata_key}) that matches the
filenames of the count files.
}
