% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AFC.R
\name{.calculate_afc}
\alias{.calculate_afc}
\title{Title}
\usage{
.calculate_afc(raw.counts = NULL, metadata = NULL, order.by.col = NULL)
}
\arguments{
\item{raw.counts}{}

\item{metadata}{}

\item{order.by.col}{}
}
\value{

}
\description{
Title
}
