% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{.calculate_pca}
\alias{.calculate_pca}
\title{Compute PCA from a c-score matrix}
\usage{
.calculate_pca(cscore.matrix = NULL)
}
\arguments{
\item{cscore.matrix}{matrix of c-score extracted from a SummarizedExperiment with ' \code{\link{extract_data}}}
}
\value{
dudi.pca object
}
\description{
Compute PCA from a c-score matrix
}
\keyword{internal}
