% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLE.R
\name{.calculate_relative_log_expression}
\alias{.calculate_relative_log_expression}
\title{Title}
\usage{
.calculate_relative_log_expression(mat = NULL, ...)
}
\arguments{
\item{mat}{}

\item{...}{}
}
\value{

}
\description{
Title
}
