% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coa.R
\name{.compute_coa}
\alias{.compute_coa}
\title{(internal) Correspondence Analysis computation function}
\usage{
.compute_coa(raw_counts = NULL)
}
\arguments{
\item{raw_counts}{A matrix of counts, as exported by extract_data()}
}
\value{
A coa object
}
\description{
(internal) Correspondence Analysis computation function
}
\keyword{internal}
