% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rlc.R
\name{.compute_RLC}
\alias{.compute_RLC}
\title{Compute Relative Log Expression (RLC)}
\usage{
.compute_RLC(mat = NULL)
}
\arguments{
\item{mat}{Data matrix to compute RLC on}
}
\value{
a RLC matrix
}
\description{
Compute Relative Log Expression (RLC)
}
\keyword{internal}
