% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rle.R
\name{.compute_RLE}
\alias{.compute_RLE}
\title{Compute Relative Log Expression (RLE)}
\usage{
.compute_RLE(mat = NULL)
}
\arguments{
\item{mat}{Data matrix to compute RLE on}
}
\value{
a RLE matrix
}
\description{
Compute Relative Log Expression (RLE)
}
\keyword{internal}
