% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{.compute_rna_cscore}
\alias{.compute_rna_cscore}
\title{compute both C-score and Z-score for one RNA of one sample.}
\usage{
.compute_rna_cscore(ds, flanking = 6, method)
}
\arguments{
\item{ds}{dataframe of a given RNA for a given sample}

\item{flanking}{the number of flanking position to use for the window}

\item{method}{either "median" or "mean", depending on how the cscore must be computed.}
}
\value{
a single RNA dataframe with cscore and flanking columns
}
\description{
compute both C-score and Z-score for one RNA of one sample.
}
\keyword{internal}
