% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{.compute_sample_cscore}
\alias{.compute_sample_cscore}
\title{compute cscore for a single sample}
\usage{
.compute_sample_cscore(sample_df = NULL, flanking = 6, method)
}
\arguments{
\item{sample_df}{the whole a dataframe for a single sample (all rRNA combined)}

\item{flanking}{the number of flanking position to use for the window}

\item{method}{either "median" or "mean", depending how the cscore must be computed.}
}
\value{
a sample dataframe with cscore and flanking columns
}
\description{
compute cscore for a single sample
}
\keyword{internal}
