% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_named_position.R
\name{.generate_riboclass_named_position}
\alias{.generate_riboclass_named_position}
\title{Generate the default name for positions
This function is used to generate the named_position column in a given sample.
It is always called when generating a RiboClass.}
\usage{
.generate_riboclass_named_position(sample_count_list, rna_col, rnapos_col)
}
\arguments{
\item{sample_count_list}{list of count dataframe (one sample = one count dataframe)}

\item{rna_col}{name or index of the column containing RNA names}

\item{rnapos_col}{name or index of the column containing position in RNA}
}
\value{
a list of sample data with the added "named_position" column
}
\description{
Generate the default name for positions
This function is used to generate the named_position column in a given sample.
It is always called when generating a RiboClass.
}
\keyword{internal}
