% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboClass.R
\name{.generate_rna_names_table}
\alias{.generate_rna_names_table}
\title{(internal) Generate a table with former and current rna names for a given RiboClass}
\usage{
.generate_rna_names_table(count_df)
}
\arguments{
\item{count_df}{a count file dataframe, containing a columns with rna names}
}
\value{
dataframe with former and current rna names
}
\description{
(internal) Generate a table with former and current rna names for a given RiboClass
}
\keyword{internal}
