% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AFC.R
\name{.plot_afc}
\alias{.plot_afc}
\title{Title}
\usage{
.plot_afc(dudi.coa = NULL, metadata = NULL, col.by.col = NULL)
}
\arguments{
\item{dudi.coa}{}

\item{metadata}{}

\item{col.by.col}{}
}
\value{

}
\description{
Title
}
