% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{.plot_boxplot}
\alias{.plot_boxplot}
\title{(iternal) plot boxplot for counts values}
\usage{
.plot_boxplot(
  matrix,
  values_col_name,
  values_to_plot,
  metadata,
  metadata_col,
  show_outlier
)
}
\arguments{
\item{matrix}{}

\item{metadata}{}

\item{metadata_col}{}

\item{show_outlier}{}
}
\value{

}
\description{
(iternal) plot boxplot for counts values
}
