% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{.plot_boxplot_samples}
\alias{.plot_boxplot_samples}
\title{(internal) plot boxplot for a given matrix of values}
\usage{
.plot_boxplot_samples(
  matrix,
  values_col_name,
  metadata,
  color_col = NA,
  outlier,
  horizontal,
  sample_colors = NULL
)
}
\arguments{
\item{matrix}{Sites x Samples C-score/count matrix (output of extract_data()).}

\item{values_col_name}{Name of the column containing the value (either count
or cscore).}

\item{metadata}{Metadata of samples in matrix.}

\item{color_col}{Name of the column in the metadata used for coloring samples.}

\item{outlier}{Show boxplot outlier values.}

\item{horizontal}{Show boxplot horizontally.}
}
\value{
ggplot boxplot
}
\description{
(internal) plot boxplot for a given matrix of values
}
\keyword{internal}
