% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{.plot_boxplot_sites}
\alias{.plot_boxplot_sites}
\title{Title}
\usage{
.plot_boxplot_sites(matrix, values_to_plot, outlier, horizontal)
}
\arguments{
\item{matrix}{}

\item{values_to_plot}{}

\item{outlier}{}

\item{values_col_name}{}
}
\description{
Title
}
