% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FCA.R
\name{.plot_fca}
\alias{.plot_fca}
\title{(internal) Plot a coa object}
\usage{
.plot_fca(dudi.coa = NULL, metadata = NULL, col.by.col = NULL)
}
\arguments{
\item{dudi.coa}{a coa object}

\item{metadata}{}

\item{col.by.col}{}
}
\value{

}
\description{
(internal) Plot a coa object
}
