% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{.plot_pca}
\alias{.plot_pca}
\title{Plot a dudi.pca object using factoextra}
\usage{
.plot_pca(
  dudi.pca = NULL,
  metadata = NULL,
  color_col = NULL,
  axes = axes,
  title = "default",
  subtitle = "samples"
)
}
\arguments{
\item{dudi.pca}{a dudi.pca object generated with Ade4}

\item{metadata}{metadata table from SummarizedExperiment}

\item{color_col}{Name of the column in the metadata used for coloring samples.}

\item{axes}{Two-element vector indicating which pair of principal components
to show.}

\item{title}{Title to display on the plot. 'default' for default title.}

\item{subtitle}{Subtitle to display on the plot. 'samples' for number of
samples. 'none' for no subtitle.}
}
\value{
a ggplot
}
\description{
Plot a dudi.pca object using factoextra
}
\keyword{internal}
