% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rlc.R
\name{.plot_rlc}
\alias{.plot_rlc}
\title{(internal) Compute and plot Relative Log Expression for a chosen data in as RiboClass}
\usage{
.plot_rlc(mat = NULL, show_outlier, col_to_plot)
}
\arguments{
\item{mat}{matrix}

\item{show_outlier}{Show outliers on the plot ?}

\item{col_to_plot}{data column to plot values from}
}
\value{
a ggplot object
}
\description{
(internal) Compute and plot Relative Log Expression for a chosen data in as RiboClass
}
\keyword{internal}
