% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_se.R
\name{.read_count_files}
\alias{.read_count_files}
\title{Import and transform a list of count files for a SummarizedExperiment}
\usage{
.read_count_files(
  path_to_files,
  sep,
  header,
  rna_col,
  position_col,
  count_value,
  metadata_filenames = NULL
)
}
\arguments{
\item{path_to_files}{path to the folder containing count files}

\item{sep}{delimiter used in count files}

\item{header}{Boolean, specify if count files have a header}

\item{rna_col}{Name or index of the column containing RNA names}

\item{position_col}{Name or index of the column containing the site positions}

\item{count_value}{Name or index of the column containing the count values}

\item{metadata_filenames}{Filenames in metadata to check against (optional)}
}
\value{
a list of sample dataframes
}
\description{
This internal function is used to read all count files and transform them into a list of dataframes.
}
\keyword{internal}
