% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ribo_count_matrix.R
\name{.update_ribo_count_with_matrix}
\alias{.update_ribo_count_with_matrix}
\title{Update count values inside a SummarizedExperiment with a matrix of values}
\usage{
.update_ribo_count_with_matrix(ribo, matrix_new)
}
\arguments{
\item{ribo}{a SummarizedExperiment object}

\item{matrix_new}{a position x sample matrix containing the new values}
}
\value{
a SummarizedExperiment with updated values
}
\description{
Update count values inside a SummarizedExperiment with a matrix of values
}
\keyword{internal}
