% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_scores_metadata.R
\name{export.results.XLS}
\alias{export.results.XLS}
\title{Export results to Excel}
\usage{
export.results.XLS(results = NULL, output.folder = "",
  select.category = "", output.file.prefix = "", digits = 5)
}
\arguments{
\item{digits:}{an interger indicating the number of digits used for rounding the numerical values of the results table}
}
\value{
Exports object to output folder
}
\description{
Export results to Excel
}
\examples{
export.results.XLS(res, select.category="Methylated")
}
