% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_scores_metadata.R
\name{general.infos.for.excel.output}
\alias{general.infos.for.excel.output}
\title{Extracts literature information from internal data sets for known/suspected methylation sites}
\usage{
general.infos.for.excel.output(df.all.samples = NULL, annot = NULL,
  select.category = "known", species = "human", annot.project.col = 1,
  annot.rna.col = 4, annot.name.col = 5, GC.min = 0.2, GC.max = 0.8,
  verbose = TRUE)
}
\arguments{
\item{df.all.samples:}{a data frame, output of select.methylated.positions()}

\item{annot:}{a data frame with general project informations. Must have columns specifying the project, rRNA, sample names, paths.}

\item{select.category:}{a character string specifying the category of methylated positions. Must be one of "known" (default), "suspected" or "candidate"}

\item{species:}{a character string specifying the species. Default is "human". In this version of the R-packge, only "human" is supported.}

\item{annot.project.col:}{a numerical value specifying the column of the project in the annotation <annot>. Default is 1.}

\item{annot.rna.col:}{a numerical value specifying the column of the rRNA in the annotation <annot>. Default is 4.}

\item{annot.name.col:}{a numerical value specifying the column of the sample name in the annotation <annot>. Default is 5.}

\item{GC.min:}{a numeric value between 0 and 1. Minimal GC-content of the surrounding 12-nt window to keep a genomic position and set GC-filter to PASS. Default is 0.2.}

\item{GC.max:}{a numeric value between 0 and 1. Maximal GC-content of the surrounding 12-nt window to keep a genomic position and set GC-filter to PASS. Default is 0.8.}

\item{verbose:}{}
}
\value{
Data frame of methylation sites, completed with information from the literature
}
\description{
Prepares the general informations (annotation) of the output for Excel export, will be completed from the corresponding reference anntation data set parameters:
}
\examples{
general.infos.for.excel.output(df.all.samples, annot, select.category="known", species="human")
}
