% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_name_positions.R
\name{generate_name_positions}
\alias{generate_name_positions}
\title{Generate "named_position" column}
\usage{
generate_name_positions(df, rna, pos)
}
\arguments{
\item{df}{dataframe with both rna and pos columns}

\item{rna}{name or position of rna column}

\item{pos}{name or position of position in rna column}
}
\value{
a dataframe
}
\description{
Generate "named_position" column
}
\keyword{internal}
