% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_palette.R
\name{generate_palette}
\alias{generate_palette}
\title{Generate the col argument for ComplexHeatmap::HeatmapAnnotation}
\usage{
generate_palette(metadata, cols_to_use)
}
\arguments{
\item{metadata}{A data.frame containing sample metadata}

\item{color_col}{The name of the column to generate colors for}
}
\value{
A named vector of colors
}
\description{
Generate the col argument for ComplexHeatmap::HeatmapAnnotation
}
