% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Title}
\usage{
generate_report(
  data_path,
  data_sep,
  metadat_path,
  metadata_sep,
  project_name = "Untitled Project",
  biological_condition_col,
  library_col
)
}
\arguments{
\item{project_name}{Name of the project}

\item{biological_condition_col}{name of the column containing the biological condition for each sample.}

\item{library_col}{name of the column containing the library for each sample.}
}
\value{

}
\description{
Title
}
