% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboClass.R
\name{generate_riboclass_named_position}
\alias{generate_riboclass_named_position}
\title{Generate the default name for positions
This function is used to generate the named_position column in a given sample.
It is always called when generating a riboClass.}
\usage{
generate_riboclass_named_position(sample_count_list, rna_col, rnapos_col)
}
\arguments{
\item{sample_count_list}{}

\item{rna_col}{}

\item{rnapos_col}{}
}
\value{

}
\description{
Generate the default name for positions
This function is used to generate the named_position column in a given sample.
It is always called when generating a riboClass.
}
