% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboClass.R
\name{generate_rna_names_table}
\alias{generate_rna_names_table}
\title{Generate a table with current rna names for a given riboClass
The table also contains}
\usage{
generate_rna_names_table(count_df)
}
\arguments{
\item{count_df}{}
}
\value{

}
\description{
Generate a table with current rna names for a given riboClass
The table also contains
}
