% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{get.file.prefix}
\alias{get.file.prefix}
\title{Creates file name prefix used for ouptut files}
\usage{
get.file.prefix(annot = NULL, annot.project.col = 1, annot.rna.col = 4)
}
\arguments{
\item{annot:}{a data frame with general project informations. Must have columns specifying the project, rRNA, sample names, paths.}

\item{annot.project.col:}{a numerical value specifying the column index of the project name in the annotation <annot>. Default is 1.}

\item{annot.rna.col:}{a numerical value specifying the column of the rRNA in the annotation <annot>. Default is 4.}
}
\value{
file prefix used for output files (plots and .csv files)
}
\description{
Returns file name prefixes as found in the annotation file (ProjectName and rRNA). File name prefix will be: RiboMethSeq_<ProjectName>_<rRNA>
}
\examples{
get.file.prefix(annotation)
}
