% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_content.R
\name{get.known.sites.info}
\alias{get.known.sites.info}
\title{Get position vector of methylated positions (internal function)}
\usage{
get.known.sites.info(annot = NULL, annot.rna.col = 4, species = "human")
}
\arguments{
\item{species}{}
}
\value{
Returns a vector of known methylated positions (if existing) of the ribosomic unit
}
\description{
Get position vector of methylated positions (internal function)
}
\examples{
get.known.sites.vec(annot)
}
