% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_content.R
\name{get.suspected.sites.info}
\alias{get.suspected.sites.info}
\title{Get vector of positions suspected to be methylated (internal function)}
\usage{
get.suspected.sites.info(annot = NULL, annot.rna.col = 4,
  species = "human")
}
\arguments{
\item{species}{}
}
\value{
Returns positions suspected to be methylated (if existing) of the ribosomic unit
}
\description{
Get vector of positions suspected to be methylated (internal function)
}
\examples{
get.suspected.sites.info(annot)
}
