% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_2ome_summary.R
\name{get_2ome_summary}
\alias{get_2ome_summary}
\title{Create a metric summary table for 2ome_analysis report}
\usage{
get_2ome_summary(ribo = ribo)
}
\arguments{
\item{ribo}{a RiboClass object}
}
\value{
a data frame
}
\description{
Create a metric summary table for 2ome_analysis report
}
\examples{
data("ribo_toy")
data("human_methylated")
ribo_toy <- remove_ribo_samples(ribo_toy,c("RNA1", "RNA2"))
ribo_toy <- rename_rna(ribo_toy)  
ribo_toy <- annotate_site(ribo_toy,
                                annot = human_methylated,
                                anno_rna = "rRNA",
                                anno_pos = "Position",
                                anno_value = "Nomenclature")
get_2ome_summary(ribo = ribo_toy)
}
