% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff_sites_summary.R
\name{get_diff_sites_summary}
\alias{get_diff_sites_summary}
\title{Get a summary of statistics for differential 2'ome-sites analysis}
\usage{
get_diff_sites_summary(
  ribo = ribo,
  pthr = 0.05,
  condition_col = condition_col,
  cscore_cutoff = 0.05,
  comparisons = comparisons
)
}
\arguments{
\item{ribo}{a RiboClass}

\item{pthr}{p-value threshold}

\item{condition_col}{the condition column used for calculation in metadata}

\item{cscore_cutoff}{cutoff of c-score under which sites are not taken in account}

\item{comparisons}{table given by the user}
}
\value{
a data frame
}
\description{
Get a summary of statistics for differential 2'ome-sites analysis
}
\examples{
data("ribo_toy")
data("human_methylated")
ribo_toy <- remove_ribo_samples(ribo_toy,c("RNA1", "RNA2"))
ribo_toy <- rename_rna(ribo_toy)  
ribo_toy <- annotate_site(ribo_toy,
                                annot = human_methylated,
                                anno_rna = "rRNA",
                                anno_pos = "Position",
                                anno_value = "Nomenclature")
comparisons <- tibble::tibble(
                              comp = c("comp1"),
                              ctrl = c("cond1"),
                              cases = c("cond2") )                                
get_diff_sites_summary(ribo = ribo_toy, pthr = 0.05, condition_col = "condition",
 cscore_cutoff = 0.05, comparisons = comparisons)
}
