% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IQR.R
\name{get_IQR}
\alias{get_IQR}
\title{Compute IQR or variance for 2'Ome sites}
\usage{
get_IQR(df = NULL, order = "IQR")
}
\arguments{
\item{df}{a dataframe}

\item{order}{athe type of variance to be computed. Should be var or iqr.}
}
\value{
a ggplot data frame with Cscore, Site, IQR or var value and sample ID
}
\description{
Compute IQR or variance for 2'Ome sites
}
\examples{
data("ribo_toy")
ribo_matrix <- extract_data(ribo_toy, position_to_rownames = TRUE)
ribo_matrix <- na.omit(ribo_matrix)
IQR_df <- get_IQR(ribo_matrix)
}
