% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers.R
\name{get_outliers}
\alias{get_outliers}
\title{Create a outlier summary table (used in plot_heatmap_annotated)}
\usage{
get_outliers(ribo = ribo)
}
\arguments{
\item{ribo}{a RiboClass object}
}
\value{
a data frame
}
\description{
Create a outlier summary table (used in plot_heatmap_annotated)
}
\examples{
data("ribo_toy")
get_outliers(ribo = ribo_toy)
}
