% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_test.R
\name{get_range}
\alias{get_range}
\title{Return the range between conditions' mean}
\usage{
get_range(cscore_matrix = NULL, metadata = NULL, factor_column = NULL)
}
\arguments{
\item{cscore_matrix}{Sites x Samples C-score matrix (output of extract_data()).}

\item{metadata}{Metadata associated to samples c-score matrix.}

\item{factor_column}{Metadata column used to group samples by.}
}
\value{
a dataframe with 1) annotated site name 2) range of mean between conditions for this site.
}
\description{
Return the range between conditions' mean
}
\keyword{internal}
