% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_annotated.R
\name{heatmap_annotated}
\alias{heatmap_annotated}
\title{Plot a correlation heatmap from a RiboClass object.
 
Shows the correlation distance between samples.}
\usage{
heatmap_annotated(ribo = ribo, qcdata = qcdata)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{qcdata}{a dataframe with metadata and outliers information}
}
\value{
ComplexHeatmap object
}
\description{
Plot a correlation heatmap from a RiboClass object.
 
Shows the correlation distance between samples.
}
\examples{
plot_heatmap_annotated(ribo, qcdata)
}
