% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_methylated}
\alias{human_methylated}
\title{Known 2'OMe positions in Humans' rRNA}
\format{
A data frame with 112 rows and 10 variables:
\describe{
  \item{Position}{nucleotide position on rRNA}
  \item{rRNA}{rRNA where this methylation is found}
  \item{Nomenclature}{name given to this modification}
  \item{NR_046235.Numbering}{Position on}
  \item{Sequence}{nucleotides sequence surround position}
  \item{SNORD}{ID of associated snoRNA}
  \item{Mode.of.coding}{how the SNORD is coded}
  \item{SNORD.host.gene}{SNORD's host gene}
  \item{Ensembl}{SNORD's Ensembl reference}
  \item{Nucleotide}{the nucleotide present at the position (A,T,G or C)}
  ...
}
}
\usage{
human_methylated
}
\description{
A dataset containing 112 2'OMe positions with associated data
}
\keyword{datasets}
