% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_structure.R
\name{identify_rRNA_chains}
\alias{identify_rRNA_chains}
\title{Quick function to identify rRNA chains in a ribosome structure}
\usage{
identify_rRNA_chains(pdb_id)
}
\arguments{
\item{pdb_id}{PDB ID to inspect}
}
\value{
Named vector of rRNA chain assignments
}
\description{
Quick function to identify rRNA chains in a ribosome structure
}
