% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_samples.R
\name{keep_samples}
\alias{keep_samples}
\title{Keep only selected samples from a SummarizedExperiment Object.}
\usage{
keep_samples(ribo = NULL, samples_to_keep = NULL)
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{samples_to_keep}{A vector containing the names of samples to keep.}
}
\value{
A SummarizedExperiment with only selected samples.
}
\description{
Keep only selected samples from a SummarizedExperiment Object.
}
\examples{
data("ribo_toy")
# ribo_test <- keep_samples(ribo = ribo_toy, samples_to_keep = c("S1", "S2"))

}
