% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_annotation.R
\name{keep_selected_annotation}
\alias{keep_selected_annotation}
\title{Keep only a subset of the current annotation}
\usage{
keep_selected_annotation(ribo, annotation_to_keep)
}
\arguments{
\item{ribo}{a SummarizedExperiment}

\item{annotation_to_keep}{vector containing annotated sites'name to keep}
}
\value{
a SummarizedExperiment where only annotated sites within annotation_to_keep are
still annotated.
}
\description{
Keep only a subset of the current annotation
}
\examples{
data("ribo_toy")
data("human_methylated")
# ribo_toy <- rename_rna(ribo_toy)
# ribo_toy <- annotate_site(ribo_toy,human_methylated)
# ribo_toy <- keep_selected_annotation(ribo_toy, c("28S_Am1310","28S_Cm2848"))

}
