% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_test.R
\name{kruskal_test_on_cscores}
\alias{kruskal_test_on_cscores}
\title{Kruskal test on c-scores}
\usage{
kruskal_test_on_cscores(
  cscore_matrix = NULL,
  metadata = NULL,
  adjust_pvalues_method = "fdr",
  factor_column = NULL,
  statistical_test = "kruskal"
)
}
\arguments{
\item{cscore_matrix}{Sites x Samples C-score matrix (output of extract_data()).}

\item{metadata}{Metadata associated to samples c-score matrix.}

\item{adjust_pvalues_method}{Method to adjust p-value, one of p.adjust.methods.}

\item{factor_column}{Metadata column used to group samples by.}

\item{statistical_test}{Statistical test used to compute p-values.
One of "kruskal" (wilcox is automatically used if there are 2 groups) or "t.test".}
}
\value{
A dataframe with three columns : 1) site : RNA site, 2) p.val : raw p-values, 3) adjusted p-values.
}
\description{
Kruskal test on c-scores
}
\keyword{internal}
