% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_content.R
\name{load.gc.content.for.human.ribosome}
\alias{load.gc.content.for.human.ribosome}
\title{Load GC content and positions of methylated sites for human ribosomes}
\usage{
load.gc.content.for.human.ribosome(annot = NULL, annot.rna.col = 4)
}
\arguments{
\item{annot.rna.col}{}
}
\value{
List of GC content of known methylated and suspected positions
}
\description{
Load GC content and positions of methylated sites for human ribosomes
}
\examples{
load.gc.content.for.human.ribosome(annot)
}
