% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_scores_metadata.R
\name{mann.whitney.test}
\alias{mann.whitney.test}
\title{Nonparametric Mann-Whitney U-Test to compare methylation between different biological conditions}
\usage{
mann.whitney.test(output.final = NULL, annot = NULL, annot.group.col = 3,
  annot.name.col = 5, score.to.test = "both", verbose = TRUE,
  digits = 5)
}
\arguments{
\item{annot:}{a data frame with general project informations. Must have columns specifying the project, rRNA, sample names, paths.}

\item{annot.group.col:}{a numerical value specifying the column of the biological condition in the annotation <annot>. Default is 3. Note that only 2 different biological conditions are allowed.}

\item{annot.name.col:}{a numerical value specifying the column of the sample name in the annotation <annot>. Default is 5.}

\item{score.to.test:}{a character string specifying the score to which the Mann-Whitney test is applied. Must be one of "C-Score", "Z-Score", or "both".}

\item{digits:}{an integer value specifying the number of digits for rounding}
}
\value{
Result of statistical test (Nonparametric Mann-Whitney U-Test) for differential methylation between 2 biological conditions
}
\description{
Nonparametric Mann-Whitney U-Test to compare methylation between different biological conditions
}
\examples{
mann.whitney.test(output.final, annot)
}
