% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_test.R
\name{mean_max_difference}
\alias{mean_max_difference}
\title{Return the mean}
\usage{
mean_max_difference(
  cscore_matrix = NULL,
  metadata = NULL,
  factor_column = NULL
)
}
\arguments{
\item{cscore_matrix}{}

\item{metadata}{}

\item{factor_column}{}

\item{order_by_col}{}
}
\description{
Return the mean
}
\keyword{internal}
