% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{mean_samples_by_conditon}
\alias{mean_samples_by_conditon}
\title{Regroup samples by condition and calculate mean for each condition}
\usage{
mean_samples_by_conditon(ribo, value, metadata_condition)
}
\arguments{
\item{ribo}{a riboClass object}

\item{value}{value to calculate mean by condition}

\item{metadata_condition}{condition to group samples by}
}
\value{

}
\description{
Regroup samples by condition and calculate mean for each condition
}
