% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plots_ggplot.R
\name{plot.cscore.vs.zscore.gg}
\alias{plot.cscore.vs.zscore.gg}
\title{C-score versus Z-score}
\usage{
\method{plot}{cscore.vs.zscore.gg}(
  wdw.df.list = NULL,
  save.as.png = TRUE,
  display.plot = TRUE,
  output.folder = "",
  plot.prefix = "",
  xyplot.rows = 2,
  zscore.col.name = "dist2medInMad",
  cscore.col.name = "ScoreC.Median.net",
  meth.col.name = "meth",
  suspected.col.name = "suspected",
  verbose = TRUE,
  zscore.threshold = 2,
  cscore.threshold = 0.8
)
}
\arguments{
\item{save.as.png:}{a logical indicating if the graphical output should be saved as .png or not. Default is TRUE.}

\item{display.plot:}{a logical indicating if the graphical output should be displayed at the default screen device. Default is TRUE.}

\item{output.folder:}{a character string indicating a path to an existing folder where plots are saved. Default is the working directory.}

\item{plot.prefix:}{a character string used to construct the output file names. Note that the output file will start with this string.}

\item{zscore.threshold:}{a numeric value indicating the Z-score threshold for the plot, which is represented by a dotted line. Default is 2.}

\item{cscore.threshold:}{a numeric value indicating the C-score threshold for the plot, which is represented by a dotted line. Default is 0.8.}
}
\value{
plot
}
\description{
C-score versus Z-score
}
\examples{
plot.cscore.vs.zscore.gg(wdw.df.list)
}
