% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plots_ggplot.R
\name{plot.cscores.boxplot}
\alias{plot.cscores.boxplot}
\title{Boxplot of C-scores across biological conditions}
\usage{
\method{plot}{cscores.boxplot}(df = NULL, save.as.png = TRUE,
  display.plot = TRUE, output.folder = "",
  plot.analysis.type = "UNDEFINED", plot.new.candidates = FALSE,
  plot.prefix = "", verbose = TRUE)
}
\arguments{
\item{verbose}{}

\item{df:}{a data frame to be plotted. Data frame must have the following columns: position, type, identity, ScoreC.Median.net, cond, sample, locus.}

\item{save.as.png:}{a logical indicating if the graphical output should be saved as .png or not. Default is TRUE.}

\item{display.plot:}{a logical indicating if the graphical output should be displayed at the default screen device. Default is TRUE.}

\item{output.folder:}{a character string indicating a path to an existing folder where plots are saved. Default is the working directory.}

\item{plot.analysis.type:}{a character string used to construct the output file name. Suggestions: "Methylated", "Suspected", "Candidates". Default is "UNDEFINED".}

\item{plot.new.candidates:}{a logical (default is FALSE) indicating if plot is created for data from known/suspected methylation sites or to plot new candidate sites in order to define the label of facets. The identity column is used for known/suspected sites. The rRNA_position for new candidate sites.}

\item{plot.prefix:}{a character string used to construct the output file names. Note that the output file will start with this string.}
}
\value{
Boxplot of C-scores by different biological groups. Each genomic positions ons appear as different panels. Samples are diplayed by different colors.
}
\description{
Boxplot of C-scores across biological conditions
}
\examples{
plot.cscores.boxplot(df)
}
