% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots_ggplot.R
\name{plot.quality.control.correlation}
\alias{plot.quality.control.correlation}
\title{Quality control plot}
\usage{
\method{plot}{quality.control.correlation}(
  all.samples = NULL,
  output.folder = "",
  plot.prefix = "",
  save.as.png = TRUE,
  display.plot = TRUE,
  data.counts.col = 4,
  data.position.col = 1,
  verbose = TRUE
)
}
\arguments{
\item{output.folder:}{a character string indicating the path to output folder where the plots will be saved. Default is the working directory.}

\item{plot.prefix:}{a character string indicating the prefix of the file name that is used when the plot is saved as .png. Default is the empty string.}

\item{save.as.png:}{a logical indicating if the graphical output should be saved as .png or not. Default is TRUE.}

\item{display.plot:}{a logical indicating if the graphical output should be displayed at the default screen device. Default is TRUE.}

\item{data.counts.col:}{a numeric value specifying the index of the column of the data matrix with the counts. Default is 4.}

\item{data.position.col:}{a numeric value specifying the index of the column of the data matrix with the genomic position. Default is 1.}

\item{verbose:}{a logical indicating if the ouput messages will be printed to the console. Default is TRUE.}
}
\value{
plot
}
\description{
Plots the correlation of read counts between all samples
}
\examples{
plot.quality.control.correlation(all.samples)
}
