% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plots_ggplot.R
\name{plot.volcano.of.scores}
\alias{plot.volcano.of.scores}
\title{Volcano plot of C-Scores}
\usage{
plot.volcano.of.scores
}
\arguments{
\item{score.to.test:}{a string indicating the type of score to use for the volcano plot. Can by <C-Score> or <Z-Score>. Default is <C-Score>.}

\item{save.as.png:}{a logical indicating if the graphical output should be saved as .png or not. Default is TRUE.}

\item{output.folder:}{a character string indicating a path to an existing folder where plots are saved. Default is the working directory.}

\item{plot.prefix:}{a character string used to construct the output file names. Note that the output file will start with this string.}

\item{plot.analysis.type:}{a character string used to construct the output file name. Suggestions: "Methylated", "Suspected", "Candidates". Default is "UNDEFINED".}

\item{display.plot:}{a logical indicating if the graphical output should be displayed at the default screen device. Default is TRUE.}

\item{xintercept.neg:}{a numerical value indicating an x-value where to add a dotted vertical line at the volcano plot.}

\item{xintercept.pos:}{a numerical value indicating an x-value where to add a dotted vertical line at the volcano plot.}

\item{y.intercept:}{a numerical value indicating an y-value where to add a dotted horizontal line at the volcano plot. All points beyond will be displayed with their labels}
}
\value{
Volcano plot (png) of scores, using fold change and difference on x-axis
}
\description{
Volcano plot of C-Scores
}
\examples{
plot.volcano.of.scores(result.statistical.test, output.folder)
}
