% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AFC.R
\name{plot_AFC}
\alias{plot_AFC}
\title{Title}
\usage{
plot_AFC(ribo, col_to_plot, order_by_col, col_for_color = NULL)
}
\arguments{
\item{ribo}{}

\item{col_to_plot}{}

\item{order_by_col}{}

\item{col_for_color}{}
}
\value{

}
\description{
Title
}
