% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barpos.R
\name{plot_barpos}
\alias{plot_barpos}
\title{Title}
\usage{
plot_barpos(ribo, values_col, metadata_col)
}
\arguments{
\item{ribo}{}

\item{values_col}{}

\item{metadata_col}{}
}
\value{

}
\description{
Title
}
