% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coa.R
\name{plot_coa}
\alias{plot_coa}
\title{Correspondence analysis of a SummarizedExperiment' counts.}
\usage{
plot_coa(
  ribo,
  color_col = NULL,
  axes = c(1, 2),
  only_annotated = FALSE,
  title = "default",
  subtitle = "default",
  object_only = FALSE
)
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{color_col}{Name of the column in the metadata used for coloring samples.}

\item{axes}{Two-element vector indicating which pair of COA components
to show.}

\item{only_annotated}{If TRUE, use only annotated sites to plot COA.}

\item{title}{Title to display on the plot. 'default' for default title.}

\item{subtitle}{Subtitle to display on the plot. 'samples' for number of
samples. 'none' for no subtitle.}

\item{object_only}{Return directly the full dudi.coa object, without
generating the plot.}
}
\value{
A ggplot or a dudi.coa object if object_only is set to True.
}
\description{
Correspondence analysis of a SummarizedExperiment' counts.
}
\examples{
data("ribo_toy")
# plot_coa(ribo = ribo_toy, color_col = 'condition')
}
