% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison_median.R
\name{plot_comparison_median}
\alias{plot_comparison_median}
\title{Plot comparison of median of c-score between two conditions}
\usage{
plot_comparison_median(
  ribo = NULL,
  condition_col = NULL,
  ech = NULL,
  ctrl = NULL
)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{condition_col}{condition column for the plot}

\item{ech}{modality 1 of condition_col}

\item{ctrl}{modality 2 of condition_col}
}
\value{
two ggplot objects
}
\description{
Plot comparison of median of c-score between two conditions
}
\examples{
plot_comparison_median(ribo_adj_annot, params$condition_col, params$ech, params$ctrl)
}
