% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC_profile.R
\name{plot_count_profile_by_cond}
\alias{plot_count_profile_by_cond}
\title{Title}
\usage{
plot_count_profile_by_cond(ribo, metadata_col)
}
\arguments{
\item{ribo}{}

\item{metadata_col}{}
}
\value{

}
\description{
Title
}
