% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FCA.R
\name{plot_fca}
\alias{plot_fca}
\title{factorial correspondence analysis of a riboclass for a given column}
\usage{
plot_fca(ribo, col_for_color = NULL, axis = c(1, 2))
}
\arguments{
\item{ribo}{A RiboClass object}

\item{col_for_color}{metadata column to colorize samples}

\item{axis}{COA axis to plot. First and second axis will be plotted by default}
}
\value{
a FCA plot
}
\description{
factorial correspondence analysis of a riboclass for a given column
}
\examples{
plot_fca(ribo = ribo, col_for_color = "condition")
}
