% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_global_profile.R
\name{plot_global_profile}
\alias{plot_global_profile}
\title{Plot global profile for a given condition on a given metadata}
\usage{
plot_global_profile(
  ribo = ribo_adj_annot,
  condition_col = condition_col,
  ech = ech,
  ctrl = ctrl
)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{condition_col}{condition column for the plot}

\item{ech}{modality 1 of condition_col}

\item{ctrl}{modality 2 of condition_col}
}
\value{
a ggplot object
}
\description{
Plot global profile for a given condition on a given metadata
}
\examples{
plot_global_profile(ribo, condition_col, ech, ctrl)
}
