% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_group_scales}
\alias{plot_group_scales}
\title{Build a consistent set of discrete ggplot scales.}
\usage{
plot_group_scales(sample_colors = NULL, color = TRUE, fill = TRUE)
}
\arguments{
\item{sample_colors}{Optional named character vector of colors.}

\item{color}{Whether to add a color scale.}

\item{fill}{Whether to add a fill scale.}
}
\description{
Build a consistent set of discrete ggplot scales.
}
\keyword{internal}
