% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cond_lineplot.R
\name{plot_lineplot}
\alias{plot_lineplot}
\title{Title}
\usage{
plot_lineplot(ribo, value_col, metadata_col)
}
\arguments{
\item{ribo}{}

\item{value_col}{}

\item{metadata_col}{}
}
\value{

}
\description{
Title
}
