% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kruskhal_test.R
\name{plot_most_significant_sites}
\alias{plot_most_significant_sites}
\title{Title}
\usage{
plot_most_significant_sites(
  df_of_Cscores = NULL,
  df_of_kruskal = NULL,
  most_significant_sites = most_significant_sites
)
}
\arguments{
\item{df_of_Cscores}{}

\item{df_of_kruskal}{}

\item{most_significant_sites}{}
}
\value{

}
\description{
Title
}
