% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rlc.R
\name{plot_rlc}
\alias{plot_rlc}
\title{Relative Log Expression plot of a RiboClass object’s counts.}
\usage{
plot_rlc(ribo, show_outlier = TRUE)
}
\arguments{
\item{ribo}{a RiboClass object.}

\item{show_outlier}{Show boxplot outlier values.}
}
\value{
A ggplot object. Samples with a median lower than median(RLC counts)-2*MAD (Median Absolute Deviation) are colored in red.
}
\description{
Relative Log Expression plot of a RiboClass object’s counts.
}
\examples{
data("ribo_toy")
plot_rlc(ribo_toy)
}
