% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites_by_iqr.R
\name{plot_sites_by_IQR}
\alias{plot_sites_by_IQR}
\title{Plot IQR or variance per site}
\usage{
plot_sites_by_IQR(ribo = NULL, plot = "IQR", variance = "IQR", ...)
}
\arguments{
\item{ribo}{A ribo class object containing the data}

\item{plot}{Type of plot: "IQR" (default) or "boxplot"}

\item{variance}{Metric to visualize: "IQR" (default) or "var" (variance)}

\item{...}{Additional theme arguments for ggplot}
}
\value{
A ggplot object
}
\description{
The sites that show an IQR/variance higher than median + 2*MAD are colored in red and are considered as the most variant sites.
}
\examples{
data("ribo_toy")
data("human_methylated")
ribo_toy <- rename_rna(ribo_toy)
ribo_toy <- annotate_site(ribo_toy, human_methylated)
plot_sites_by_IQR(ribo = ribo_toy, plot = "IQR")
}
