% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat.R
\name{plot_stat}
\alias{plot_stat}
\title{Plot statitic}
\usage{
plot_stat(ribo = ribo_adj_annot, site = NULL, res_pv = NULL, pthr = NULL)
}
\arguments{
\item{ribo}{a RiboClass}

\item{site}{sites to plot}

\item{res_pv}{df of p values extracted from res_pv.R}

\item{pthr}{p value threshold}
}
\value{
a boxplot
}
\description{
Plot statitic
}
\examples{
plot_stat(data = ribo, site = NULL, res_pv = res_pv, pthr = 0.05)
}
