% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites.R
\name{plot_test_wrapper}
\alias{plot_test_wrapper}
\title{plot interesting sites between conditions}
\usage{
plot_test_wrapper(
  ribo,
  values_column,
  factor_column,
  p_cutoff = 0.01,
  cscore_cutoff = 0.05
)
}
\arguments{
\item{ribo}{}

\item{values_column}{}

\item{factor_column}{}
}
\description{
plot interesting sites between conditions
}
