% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot a volcano plot of differential sites.}
\usage{
plot_volcano(
  ribo,
  factor_column,
  reference_level = NULL,
  target_condition = NULL,
  p_cutoff = 0.01,
  cscore_cutoff = 0.05,
  statistical_test = "kruskal",
  adjust_pvalues_method = "fdr",
  y_axis_p = "p.adj",
  object_only = FALSE
)
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{factor_column}{Metadata column used to group samples by.}

\item{reference_level}{The level in \code{factor_column} to use as the reference for calculating the Delta C-score. If NULL, the first level is used.}

\item{target_condition}{The target level in \code{factor_column} to compare against the reference. If NULL, all levels other than the reference are used.}

\item{p_cutoff}{Cutoff for the adjusted p-value (or raw p-value depending on \code{y_axis_p}) to draw a horizontal threshold line.}

\item{cscore_cutoff}{Cutoff for the absolute Delta C-score (mean difference) to draw vertical threshold lines.}

\item{statistical_test}{Statistical test used to compute p-values.
One of "kruskal" (wilcox is automatically used if there are 2 groups) or "t.test".}

\item{adjust_pvalues_method}{Method used to adjust p-value (one of p.adjust.methods).}

\item{y_axis_p}{Which p-value to use for the y-axis. Either "p.adj" or "p.val".}

\item{object_only}{Return the results of the statistical test and C-score mean range in a dataframe directly, without plotting.}
}
\value{
a ggplot object or a dataframe if \code{object_only} is TRUE.
}
\description{
Display a volcano plot of the differential sites finding, with the Delta C-score (or log2 fold change)
on the x-axis and the -log10 of the p-value on the y-axis.
}
\details{
To be considered as differential, a site must follow two conditions :
\itemize{
\item Have a significant p-value on the selected statistical
test on c-score between conditions.
\item Have a absolute Delta C-score between a condition and the reference
above a certain cutoff.
}

Both the p-value cutoff and c-score range cutoff can be changed with
\code{p_cutoff} and \code{cscore_cutoff} parameters respectively.
}
\examples{
data("ribo_toy")
data("human_methylated")
# ribo_toy <- rename_rna(ribo_toy)
# ribo_toy <- annotate_site(ribo_toy,human_methylated)
# plot_volcano(ribo_toy, "condition", reference_level = "RNA ref", p_cutoff=0.1)
}
