% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_scores_metadata.R
\name{prepare.output.data.Scores}
\alias{prepare.output.data.Scores}
\alias{prepare.output.data.Scores}
\title{Prepares the output data (Cscore and Zscore) for Excel}
\usage{
prepare.output.data.Scores(df.all.samples = NULL, annot = NULL,
  annot.name.col = 5)

prepare.output.data.Scores(df.all.samples = NULL, annot = NULL,
  annot.name.col = 5)
}
\arguments{
\item{annot:}{a data frame with general project informations. Must have columns specifying the project, rRNA, sample names, paths.}

\item{annot.name.col:}{a numerical value specifying the column of the sample name in the annotation <annot>. Default is 5.}

\item{annot:}{a data frame with general project informations. Must have columns specifying the project, rRNA, sample names, paths.}

\item{annot.name.col:}{a numerical value specifying the column of the sample name in the annotation <annot>. Default is 5.}
}
\value{
Returns data frame with 1st column being the position, and two columns for each sample in <df.all.samples> corresponding to the C- and Z-Scores

data frame with 1st column = position and for each sample in df.all.samples 2 columns corresponding to the CScore and to the ZScore
}
\description{
Prepares the output data (Cscore and Zscore) for Excel

Prepares the output data (C-score and Z-score) for Excel output
}
\examples{
prepare.output.data.Scores(df.all.samples, annot)
prepare.output.data.Scores(df.all.samples, annot)
}
