% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyNum.R
\name{prettyNum}
\alias{prettyNum}
\title{prettyNum for matrices and tables}
\usage{
prettyNum(x, ...)
}
\arguments{
\item{x}{a \code{data.frame}, \code{matrix} or \code{vector}}

\item{\dots}{arguments passed to \code{\link[base]{prettyNum}}}
}
\value{
A character object of same size and attributes as \code{x}, in the
    current locale's encoding.
}
\description{
Extension of the excellent \code{\link[base]{prettyNum}} from \code{base} to
 work with \code{matrix} & \code{data.frame} objects.
}
\details{
\code{prettyNum} is the utility function for prettifying \code{x}.
\code{x} can be complex (or \code{format(<complex>)}, \code{matrix} or \code{data.frame} here.
If each element of \code{x} is not a character, \code{format(x[i], ...)} 
is applied to each element, and
then it is left unchanged if all the other arguments are at their
defaults.  Note that \code{prettyNum(x)} may behave unexpectedly if elements of \code{x}
are a \code{character} vector not resulting from something like
\code{format(<number>)}: in particular it assumes that a period is a
decimal mark.
}
\seealso{
\code{\link[base]{prettyNum}}
}
\author{
Mark Cowley
}
