% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboclass.R
\name{print.RiboClass}
\alias{print.RiboClass}
\title{print() to display basic informations about a RiboClass}
\usage{
\method{print}{RiboClass}(x, ...)
}
\arguments{
\item{x}{a RiboClass}

\item{...}{print() params}
}
\value{
text
}
\description{
print() to display basic informations about a RiboClass
}
\keyword{internal}
