% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{read.and.annotate.data}
\alias{read.and.annotate.data}
\title{Reads ribomethseq sequencing files}
\usage{
read.and.annotate.data(
  data.folder = NULL,
  annot = NULL,
  species = "human",
  annot.path.col = 2,
  annot.group.col = 3,
  annot.rna.col = 4,
  annot.name.col = 5,
  data.position.col = 1,
  data.type.col = 2,
  data.identity.col = 3,
  data.counts.col = 4,
  sep = "",
  um14 = 15
)
}
\arguments{
\item{data.folder}{a character string indicating the data folder with the input files}

\item{annot}{a data frame with general project informations. Must have columns specifying the project, rRNA, sample names, paths.}

\item{species}{a character string specifying the species. Default is "human". In this version of the R-packge, only "human" is supported.}

\item{annot.path.col}{a numerical value specifying the column with the file name in the annotation <annot>. Default is 2.}

\item{annot.group.col}{a numerical value specifying the column of the biological condition in the annotation <annot>. Default is 3. Note that only 2 different biological conditions are allowed.}

\item{annot.rna.col}{a numerical value specifying the column of the rRNA in the annotation <annot>. Default is 4.}

\item{annot.name.col}{a numerical value specifying the column of the sample name in the annotation <annot>. Default is 5.}

\item{data.position.col}{a numerical value specifying the column with the genomic position of the sequencing data in the data file. Default is 1.}

\item{data.type.col}{a numerical value specifying the column with the RNA type of the sequencing data in the data file. Default is 2. If there is no column specifiying the RNA type, the parameter should set to 1.}

\item{data.identity.col}{a numerical value specifying the column with the annotation of the RNA methylation site. Default is 3. If there is no column with the annotations, the parameter should set to 1.}

\item{data.counts.col}{a numerical value specifying the column of the rRNA in the annotation <annot>. Default is 4.}

\item{sep}{the field separator character for the internal function read.table. Values on each line of the file are separated by this character. If sep is "" (the default for read.table) the separator is "white space", that is one or more spaces, tabs, newlines or carriage returns.}

\item{um14}{}
}
\value{
List of data frames. Each element is a count table of one sample, the name of the list element corresponds to the sample name
}
\description{
Reads ribomethseq sequencing files
}
\examples{
read.and.annotate.data(input.data.folder, project.annotation)
}
