% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboClass.R
\name{read_count_files}
\alias{read_count_files}
\title{Import and transform a list of count files for a riboClass}
\usage{
read_count_files(path_to_files, sep, header, rna_col, position_col, count_col)
}
\arguments{
\item{path_to_files}{path to the folder containing count files}

\item{sep}{CSV seperator}

\item{rna_col}{position of the column containing RNA name}

\item{position_col}{position of the column containing the genomic position}

\item{count_col}{position of the column contaning the count for the given genomic position}
}
\value{

}
\description{
This function should not be used by itself. It is better to generate a full riboClass.
}
