% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{read_counts}
\alias{read_counts}
\title{Read ribomethseq count files and their associated metadata}
\usage{
read_counts(
  counts_path,
  metadata_path = NA,
  rna_names,
  count_sep = "\\t",
  metadata_sep = ";",
  metadata_filename_col = 1,
  metadata_sample_name_col = 2
)
}
\arguments{
\item{counts_path:}{path to the data folder with the count files}

\item{metadata_path:}{path to the CSV file containing metadata}

\item{rna_names:}{names of the RNA used in this run}

\item{count_sep:}{delimiter used in genomecov csv}

\item{metadata_sep:}{delimiter used in metadata csv}

\item{metadata_filename_col:}{name or position of the column containing the filename}

\item{metadata_sample_name_col:}{name or position of the column containing sample name}
}
\value{
a riboclass
}
\description{
Read ribomethseq count files and their associated metadata
}
