% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_or_remove_samples_ribo.R
\name{remove_ribo_samples}
\alias{remove_ribo_samples}
\title{Remove samples from a Riboclass object}
\usage{
remove_ribo_samples(ribo = NULL, samples_to_delete = NULL)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{samples_to_delete}{A vector containing the names of samples to remove.}
}
\value{
A RiboClass without the removed samples in data and metadata.
}
\description{
Remove samples from a Riboclass object
}
\examples{
data("ribo_toy")
ribo_test <- remove_ribo_samples(ribo = ribo_toy, samples_to_delete = c("S1", "S2"))
}
