% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rna.R
\name{remove_rna}
\alias{remove_rna}
\title{Remove one or more RNA among your SummarizedExperiment' samples}
\usage{
remove_rna(ribo, rna_to_remove)
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{rna_to_remove}{One RNA name or a vector of RNA names.}
}
\value{
A SummarizedExperiment with the RNA(s) removed.
}
\description{
Remove one or more RNA among your SummarizedExperiment' samples
}
\examples{
data("ribo_toy")
# ribo_toy <- remove_rna(ribo_toy, "NR_023363.1_5S")
}
