% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_samples.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples from a SummarizedExperiment object}
\usage{
remove_samples(ribo = NULL, samples_to_delete = NULL)
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{samples_to_delete}{A vector containing the names of samples to remove.}
}
\value{
A SummarizedExperiment without the removed samples.
}
\description{
Remove samples from a SummarizedExperiment object
}
\examples{
data("ribo_toy")
# ribo_test <- remove_samples(ribo = ribo_toy, samples_to_delete = c("S1", "S2"))

}
