% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_rna.R
\name{rename_rna}
\alias{rename_rna}
\title{Rename RNAs in a SummarizedExperiment object}
\usage{
rename_rna(ribo, new_names = c("5S", "5.8S", "18S", "28S"))
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{new_names}{Vector of new RNA names (by RNA size order).}
}
\value{
a SummarizedExperiment with updated RNA names.
}
\description{
Rename RNAs in a SummarizedExperiment object
}
\examples{
data("ribo_toy")
# ribo_toy <- rename_rna(ribo_toy ,c("5S","5.8S","18S","28S"))

}
