% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_html_wilcoxon.R
\name{render_html_wilcoxon}
\alias{render_html_wilcoxon}
\title{Print Wilcoxon result in report}
\usage{
render_html_wilcoxon(ribo, condition_col, ctrl, cases, pthr, cscore_cutoff)
}
\arguments{
\item{ribo}{a RiboClass object}

\item{condition_col}{the condition column used for calculation in metadata}

\item{ctrl}{control in comparisons table}

\item{cases}{case in comparison table}

\item{pthr}{threshold of the p_value}

\item{cscore_cutoff}{threshold of c-score}
}
\value{
a tag list
}
\description{
Print Wilcoxon result in report
}
\examples{
data("ribo_toy")
data("human_methylated")
ribo_toy <- remove_ribo_samples(ribo_toy,c("RNA1", "RNA2"))
ribo_toy <- rename_rna(ribo_toy)  
ribo_toy <- annotate_site(ribo_toy,
                                annot = human_methylated,
                                anno_rna = "rRNA",
                                anno_pos = "Position",
                                anno_value = "Nomenclature")
comparisons <- tibble::tibble(comp = c("comp1"),
                              ctrl = c("cond1"),
                              cases = c("cond2") )  
ctrl <- comparisons$ctrl
cases <- comparisons$cases 
render_html_wilcoxon(ribo_toy, "condition", ctrl, cases, pthr = 0.05, cscore_cutoff = 0.05)
}
\keyword{internal}
