% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_2ome_sites}
\alias{report_2ome_sites}
\title{Generate html report about the analysis of 2’Ome profile}
\usage{
report_2ome_sites(
  ribo,
  specie = "not specified",
  condition_col,
  project_name = "Unnamed project",
  output_dir = getwd(),
  comments = NULL
)
}
\arguments{
\item{ribo}{a RiboClass object.}

\item{specie}{studied specie of the project}

\item{condition_col}{name or index of the column __in metadata__ containing the condition}

\item{project_name}{Name of the project to display on the report}

\item{output_dir}{Path of output directory}

\item{comments}{Path to a text file containing comments}
}
\description{
Generate html report about the analysis of 2’Ome profile
}
\examples{
data('ribo_toy')
data("human_methylated")
ribo_adj_small <- remove_ribo_samples(ribo_toy, c("RNA1","RNA2"))
ribo_adj_small_name <- rename_rna(ribo_adj_small, 
new_names = c("5S", "5.8S", "18S", "28S"))
ribo_adj_annot <- annotate_site(ribo_adj_small_name, 
                                annot = human_methylated,
                                anno_rna = "rRNA",
                                anno_pos = "Position",
                                anno_value = "Nomenclature") 
report_2ome_sites(ribo = ribo_adj_annot, specie = "human", 
condition_col = "condition", project_name = "name", comments = "./comment_2ome_comp1.Rmd") 
}
