% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_diff_sites}
\alias{report_diff_sites}
\title{Generate html report about the statistical analysis of difference in 2’Ome 
level at each site}
\usage{
report_diff_sites(
  ribo,
  specie = "not specified",
  ribo_name,
  rmdfile,
  condition_col,
  library_col,
  project_name = "Unnamed project",
  output_dir = getwd(),
  comparisons = NULL,
  comments = NULL
)
}
\arguments{
\item{ribo}{a RiboClass object.}

\item{specie}{studied specie of the project}

\item{ribo_name}{name of RiboClass}

\item{rmdfile}{filename of the Rmarkdown template to use.}

\item{condition_col}{name or index of the column __in metadata__ containing the condition}

\item{library_col}{name or index of the column __in metadata__ containing the library}

\item{project_name}{Name of the project to display on the report}

\item{output_dir}{Path of output directory}

\item{comparisons}{data.frame or tibble with columns 'ctrl' and 'cases', 
defining the pairwise comparisons to run.}

\item{comments}{Path to a text file containing comments}
}
\description{
Generate html report about the statistical analysis of difference in 2’Ome 
level at each site
}
\examples{
library(dplyr)
data('ribo_toy')
data("human_methylated")
ribo_adj_small <- remove_ribo_samples(ribo_toy, c("RNA1","RNA2"))
ribo_adj_small_name <- rename_rna(ribo_adj_small, 
new_names = c("5S", "5.8S", "18S", "28S"))
ribo_adj_annot <- annotate_site(ribo_adj_small_name, 
                                annot = human_methylated,
                                anno_rna = "rRNA",
                               anno_pos = "Position",
                               anno_value = "Nomenclature")
comparisons <- tibble::tibble(
comp = c("comp1"),
ctrl = c("cond1"),
cases = c("cond2"))
kept_samples <- ribo_adj_annot$metadata \%>\% 
  dplyr::filter(!is.na(comp1)) \%>\% 
  dplyr::pull(samplename)
ribo_adj_annot_comp1 <- keep_ribo_samples(ribo_adj_annot, kept_samples)
report_diff_sites(ribo = ribo_adj_annot_comp1, specie = "human", 
condition_col = "comp1", project_name = "name", comparisons = comparisons, 
comments = "./comment_diff_site.Rmd") 
}
