% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_qc}
\alias{report_qc}
\title{Generate html report about possible batch effect for a given SummarizedExperiment.}
\usage{
report_qc(
  ribo,
  library_col,
  project_name = "Unnamed project",
  output_dir = getwd()
)
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{library_col}{Library/run column in the metadata.}

\item{project_name}{Name of the project.}

\item{output_dir}{Path to output dir. Working directory is selected by default.}
}
\description{
Generate html report about possible batch effect for a given SummarizedExperiment.
}
