% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_pv.R
\name{res_pv}
\alias{res_pv}
\title{Compute statistics for c-score based on condition column}
\usage{
res_pv(ribo = ribo_adj_annot, test = "student", condition_col = PR3)
}
\arguments{
\item{ribo}{a RiboClass object}

\item{test_name}{name of the test to perform}

\item{col_condition}{condition column on which perform the tests}
}
\value{
a dataframe
}
\description{
Compute statistics for c-score based on condition column
}
\examples{
res_pv(ribo = ribo_adj_annot, test_name = t.test, condition_col = col)
}
