% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_plot_text}
\alias{resolve_plot_text}
\title{Resolve a user-supplied plot title/subtitle value.}
\usage{
resolve_plot_text(
  value,
  default_value,
  arg_name,
  allow_none = FALSE,
  default_tokens = "default",
  none_value = ggplot2::waiver()
)
}
\arguments{
\item{value}{Input value supplied by the user.}

\item{default_value}{Default value returned when \code{value = "default"}.}

\item{arg_name}{Argument name for validation errors.}

\item{allow_none}{Whether \code{"none"} is accepted.}

\item{none_value}{Value returned when \code{value = "none"}.}
}
\description{
Resolve a user-supplied plot title/subtitle value.
}
\keyword{internal}
