% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combatSeq.R
\name{ribo_combatseq}
\alias{ribo_combatseq}
\title{Correct batch effect with ComBat seq method}
\usage{
ribo_combatseq(ribo, library_col)
}
\arguments{
\item{ribo}{a riboclass}

\item{library_col}{name of the column containing the library for each sample (metadata)}
}
\value{
a riboClass with adjusted counts
}
\description{
Correct batch effect with ComBat seq method
}
