% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ribo_toy}
\alias{ribo_toy}
\title{SummarizedExperiment from a toy dataset}
\format{
a SummarizedExperiment object with:
\describe{
  \item{assays}{list containing counts and cscore matrices}
  \item{colData}{metadata dataframe of all samples}
  \item{rowData}{dataframe containing positions, RNA names and site annotations}
  \item{metadata}{list containing the rna_names table}
  }
}
\usage{
data(ribo_toy)
}
\description{
A SummarizedExperiment object containing 10 samples + 2 reference RNA.
}
\details{
Samples are from 4 different biological conditions ("condition" column in metadata).
The sequencing has been done in two different batches ("run" column in metadata).
Both batches have the same reference RNA, to detect technical bias.
}
\keyword{datasets}
