% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_rRMSAnalyzer}
\alias{scale_fill_rRMSAnalyzer}
\title{Custom fill palette for rRMSAnalyzer}
\usage{
scale_fill_rRMSAnalyzer(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{scale_fill_manual}}.}
}
\value{
A discrete fill scale.
}
\description{
A colorblind-friendly palette for categorical variables.
}
