% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_positions_and_candidates.R
\name{select.methylated.positions}
\alias{select.methylated.positions}
\title{Selection of genomic positions}
\usage{
select.methylated.positions(
  wdw.df.list = NULL,
  annot = NULL,
  select.category = "known",
  annot.name.col = 5,
  annot.group.col = 3
)
}
\arguments{
\item{annot:}{a data frame with general project informations. Must have columns specifying the project, rRNA, sample names, paths.}

\item{select.category:}{a character string specifying the category of methylated positions. Must be one of "known" (default), "suspected" or "candidate"}

\item{annot.name.col:}{a numerical value specifying the column index of the sample name in the annotation <annot>. Default is 5.}

\item{annot.group.col:}{a numerical value specifying the column of the biological condition in the annotation <annot>. Default is 3. Note that only 2 different biological conditions are allowed.}
}
\value{
data frame of known (or suspected) methylated positions for all samples
}
\description{
Selection of either known methylated positions or suspected methylated positions from data frame (list of data frames) with C-Scores, Z-Scores, GC content information
}
\examples{
select.methylated.positions(wdw.df.list, annot, select.category="suspected")
}
