% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_positions_and_candidates.R
\name{select.new.candidates}
\alias{select.new.candidates}
\title{Select new candidates based on minimal tresholds for C-Scores and Z-Scores and within a predefined GC-content window. Already known methylated sites are filtered out, suspected sites are kept.}
\usage{
select.new.candidates(
  wdw.df.list = NULL,
  annot = NULL,
  data.position.col = 1,
  annot.name.col = 5,
  annot.group.col = 3,
  GC.min = 0.2,
  GC.max = 0.8,
  zscore = 2,
  cscore = 0.8
)
}
\arguments{
\item{annot:}{a data frame with general project informations. Must have columns specifying the project, rRNA, sample names, paths.}

\item{annot.name.col:}{a numerical value specifying the column index of the sample name in the annotation <annot>. Default is 5.}

\item{annot.group.col:}{a numerical value specifying the column index of the biological group in the annotation <annot>. Default is 3.}

\item{GC.min:}{a numeric value between 0 and 1. Minimal GC-content of the surrounding 12-nt window to keep a genomic position and set GC-filter to PASS. Default is 0.2.}

\item{GC.max:}{a numeric value between 0 and 1. Maximal GC-content of the surrounding 12-nt window to keep a genomic position and set GC-filter to PASS. Default is 0.8.}

\item{zscore:}{a numeric value above which genomic sites are kept as potential new methylation candidates. Default is 2.}

\item{cscore:}{a numeric value above which genomic sites are kept as potential new methylation candidates. Default is 0.8.}
}
\value{
Data frame of genomic positions selected to be new methylation candidates corresponding to the criterias (C- and Z-scores, GC-content). Note that positions are kept if they fulfill the 3 criteria in at least one sample of the cohort.
}
\description{
Select new candidates based on minimal tresholds for C-Scores and Z-Scores and within a predefined GC-content window. Already known methylated sites are filtered out, suspected sites are kept.
}
\examples{
select.new.candidates(wdw.df.list, annot)
}
