% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kruskhal_test.R
\name{select_most_significant_sites}
\alias{select_most_significant_sites}
\title{Title}
\usage{
select_most_significant_sites(
  df_of_kruskal = NULL,
  pCutoff = 0.01,
  CscoreCutoff = 0.05
)
}
\arguments{
\item{df_of_kruskal}{}

\item{pCutoff}{}

\item{CscoreCutoff}{}
}
\value{

}
\description{
Title
}
