% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboClass.R
\name{update_riboclass_rna_names}
\alias{update_riboclass_rna_names}
\title{Update rna names}
\usage{
update_riboclass_rna_names(ribo)
}
\arguments{
\item{rna_col}{}

\item{rna_names}{}
}
\value{

}
\description{
Update rna names
}
\examples{

}
