% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_clustering_args}
\alias{validate_clustering_args}
\title{Validate clustering arguments shared by heatmap-like plots.}
\usage{
validate_clustering_args(
  clustering_distance = "manhattan",
  clustering_method = "ward.D2"
)
}
\arguments{
\item{clustering_distance}{Distance metric name.}

\item{clustering_method}{Linkage method name.}
}
\description{
Validate clustering arguments shared by heatmap-like plots.
}
\keyword{internal}
