% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_kruskal.R
\name{wrapper.kruskal.test}
\alias{wrapper.kruskal.test}
\title{Title}
\usage{
wrapper.kruskal.test(
  ribo = NULL,
  order_by_col = NULL,
  adjust_pvalues_method = "fdr",
  factor_column = NULL
)
}
\arguments{
\item{order_by_col}{}

\item{adjust_pvalues_method}{}

\item{factor_column}{}

\item{cscore_matrix}{}

\item{metadata}{}

\item{column_or_row}{}
}
\value{

}
\description{
Title
}
