% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_test.R
\name{wrapper_kruskal_test}
\alias{wrapper_kruskal_test}
\title{(internal) Test which sites are the most variable between conditions}
\usage{
wrapper_kruskal_test(
  ribo = NULL,
  adjust_pvalues_method = "fdr",
  factor_column = NULL,
  statistical_test = "kruskal"
)
}
\arguments{
\item{ribo}{a SummarizedExperiment}

\item{adjust_pvalues_method}{p-value adjustment method (default : "fdr")}

\item{factor_column}{Metadata column used to group samples by.}

\item{statistical_test}{Statistical test used to compute p-values.
One of "kruskal" (wilcox is automatically used if there are 2 groups) or "t.test".}
}
\value{
A dataframe where each site is a row and with the following column :
 1) site : name of the site; 2) p.val : raw p-value; 3) p.adj : adjusted p-value 4) mean_max_min_difference : range between conditions of mean c-score
}
\description{
This function outputs the same result as plot_diff_sites, but as a dataframe.
It cannot be called directly, use "only_df = TRUE" in plot_diff_sites instead.
}
\details{
To be considered as differential, a site must follow two conditions :
\itemize{
\item Have a significant p-value on the selected statistical
test on c-score between conditions.
\item Have a C-score range between conditions (max median - min mea)
above a certain cutoff.
}

Both the p-value cutoff and c-score range cutoff can be changed with p_cutoff and cscore_cutoff parameters respectively.
}
\keyword{internal}
