% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_bias.R
\name{adjust_bias}
\alias{adjust_bias}
\title{Correct batch effect with ComBat-seq method}
\usage{
adjust_bias(ribo, batch, ncores = 1, ...)
}
\arguments{
\item{ribo}{a RiboClass object.}

\item{batch}{Name of the column in metadata that contains the batch number.}

\item{ncores}{Number of cores to use in case of multithreading.}

\item{...}{Parameters to pass to sva's \code{\link[sva]{ComBat_seq}} function.}
}
\value{
RiboClass with adjusted read end count values and automatically recomputed C-scores.
}
\description{
Batch effect of RiboMethSeq data can be adjusted using the ComBat-seq method. adjust_bias is a wrapper to perform ComBat-seq adjustment. 

It will return a new RiboClass with adjusted read end count values and C-scores automatically recomputed with the same setup parameters.
}
\details{
You must have a column with the batch number for each sample in your RiboClass’s metadata.
}
\examples{
data('ribo_toy')
ribo_toy_two <- keep_ribo_samples(ribo_toy,c('S1','RNA1','S7','RNA2'))
ribo_toy_adjusted <- adjust_bias(ribo_toy_two,'run') 

}
\references{
Yuqing Zhang, Giovanni Parmigiani, W Evan Johnson, ComBat-seq: batch effect adjustment for RNA-seq count data, NAR Genomics and Bioinformatics, Volume 2, Issue 3, 1 September 2020, lqaa078, https://doi.org/10.1093/nargab/lqaa078
}
