% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combatSeq.R
\name{adjust_bias}
\alias{adjust_bias}
\title{Correct batch effect with ComBatseq method}
\usage{
adjust_bias(ribo, batch, ...)
}
\arguments{
\item{ribo}{a RiboClass object, see constructor : \code{\link{create_riboclass}}}

\item{batch}{name of the column in metadata that contains the batch number.}

\item{...}{Parameters for sva's \code{\link[sva]{ComBat_seq}} function.}
}
\value{
RiboClass with adjusted values and recomputed cscore.
}
\description{
Correct batch effect with ComBatseq method
}
\details{
You must have a column with the batch number for each sample in your RiboClass’s metadata.
}
\examples{
data("ribo_toy")
ribo_toy_two <- keep_ribo_samples(ribo_toy,c("4283","RNA1","4272","RNA2"))
ribo_toy_adjusted <- adjust_bias(ribo_toy_two,"run") 

}
\references{
Yuqing Zhang, Giovanni Parmigiani, W Evan Johnson, ComBat-seq: batch effect adjustment for RNA-seq count data, NAR Genomics and Bioinformatics, Volume 2, Issue 3, 1 September 2020, lqaa078, https://doi.org/10.1093/nargab/lqaa078
}
