% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_site.R
\name{annotate_site}
\alias{annotate_site}
\title{Annotate sites according to a dataframe containing annotation}
\usage{
annotate_site(ribo, annot, anno_rna = 2, anno_pos = 1, anno_value = 3)
}
\arguments{
\item{ribo}{a riboclass object to annotate, see constructor : 
\code{\link{create_riboclass}}}

\item{annot}{The dataframe containing annotations.}

\item{anno_rna}{Name or index of the column in annot containing RNAs' name.}

\item{anno_pos}{Name or index of the column in annot containing site position inside RNA.}

\item{anno_value}{Name or index of the column in annot containing nomenclature to apply.}
}
\value{
An annotated riboClass (the site column in data should be filled with names from annot_value for known position).
}
\description{
This function will fill the "site" column in your riboclass's data with a nomenclature given in annot.
}
\examples{
data("ribo_toy")
data("human_methylated")
ribo_toy <- rename_rna(ribo_toy ,c("5S","5.8S","18S","28S"))
ribo_toy <- annotate_site(ribo_toy,human_methylated,anno_value ="Nomenclature")

}
