% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{boxplot_cscores}
\alias{boxplot_cscores}
\title{Plot boxplot of annotated sites c-score.
Sites are sorted by}
\usage{
boxplot_cscores(ribo, outlier = TRUE, horizontal = FALSE)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{outlier}{Show boxplot outlier values.}

\item{horizontal}{Show boxplot horizontally.}
}
\value{
a ggplot geom_boxplot
}
\description{
Plot boxplot of annotated sites c-score.
Sites are sorted by
}
\examples{
data("ribo_toy")
data("human_methylated")
ribo_toy <- rename_rna(ribo_toy)
ribo_toy <- annotate_site(ribo_toy,human_methylated)
boxplot_cscores(ribo_toy)

}
