% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{boxplot_cscores}
\alias{boxplot_cscores}
\title{Plot boxplot of annotated sites c-score.
Sites are sorted by}
\usage{
boxplot_cscores(ribo, outlier = TRUE, horizontal = FALSE)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{outlier}{Show boxplot outlier values.}

\item{horizontal}{Show boxplot horizontally.}
}
\description{
Plot boxplot of annotated sites c-score.
Sites are sorted by
}
