% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculate score for a whole sample cohort}
\usage{
calculate_score(
  dt = NULL,
  flanking = 6,
  data_rna_col = 1,
  data_counts_col = 3,
  use_multithreads = F
)
}
\arguments{
\item{dt}{}

\item{flanking}{}

\item{data_counts_col}{}

\item{data_position_col}{}
}
\value{

}
\description{
Calculate score for a whole sample cohort
}
