% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{calculate_score}
\alias{calculate_score}
\title{Title}
\usage{
calculate_score(
  dt = NULL,
  flanking = 6,
  data_counts_col = 4,
  data_position_col = 1
)
}
\arguments{
\item{dt}{}

\item{flanking}{}

\item{data_counts_col}{}

\item{data_position_col}{}
}
\value{

}
\description{
Title
}
