% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{calculate_score_by_RNA}
\alias{calculate_score_by_RNA}
\title{Calculate both C-score and Z-score for one RNA of one sample.}
\usage{
calculate_score_by_RNA(ds, flanking = 6, data_counts_col = 4)
}
\arguments{
\item{ds}{dataframe of a given RNA for a given sample}

\item{flanking}{the number of flanking position to use for the window}

\item{data.counts.col}{column number where counts value are stored}
}
\value{

}
\description{
Calculate both C-score and Z-score for one RNA of one sample.
}
