% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_metadata}
\alias{check_metadata}
\title{Check if a vector of strings correspond to valid metadata in a given RiboClass}
\usage{
check_metadata(ribo, metadata_name)
}
\arguments{
\item{ribo}{A RiboClass.}

\item{metadata_name}{The vector of string to check against ribo's metadata.}
}
\value{
Invisibly returns NULL if all strings are valid metadata; otherwise, 
the function stops with an error.
}
\description{
This function returns nothing if each string matches with a metadata.
It will stop and display a formatted error message if one or more strings in
the vector are not valid metadata.
}
\keyword{internal}
