% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sample}
\alias{check_sample}
\title{Check if a vector of strings correspond to valid samplenames in a given RiboClass}
\usage{
check_sample(ribo, sample_names)
}
\arguments{
\item{ribo}{A RiboClass.}

\item{sample_names}{The vector of string to check against ribo's samplenames}
}
\description{
This function returns nothing if each string matches with the samplenames
It will stop and display a formatted error message if one or more strings in
the vector are not valid samplenames
}
\keyword{internal}
