% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_positions.R
\name{check_sample_positions}
\alias{check_sample_positions}
\title{Check if two samples share the same positions}
\usage{
check_sample_positions(sample_1, sample_2, sample_1_name, sample_2_name)
}
\arguments{
\item{sample_1}{First sample to check against sample_2}

\item{sample_2}{Second sample to check against sample_1}

\item{sample_1_name}{name of sample 1}

\item{sample_2_name}{name of sample 2}
}
\value{
A boolean indicating if the two samples are identical.
}
\description{
Check if two samples share the same positions
}
\keyword{internal}
