% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FCA.R
\name{.calculate_fca}
\alias{.calculate_fca}
\title{(internal) factorial correspondence analysis calculation function}
\usage{
.calculate_fca(raw.counts = NULL)
}
\arguments{
\item{raw.counts}{}
}
\value{
a coa object
}
\description{
(internal) factorial correspondence analysis calculation function
}
\examples{
.calculate_fca(raw_counts = cscore_matrix)
}
