% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{.calculate_pca}
\alias{.calculate_pca}
\title{Title}
\usage{
.calculate_pca(cscore.matrix = NULL, metadata = NULL, order.by.col = NULL)
}
\arguments{
\item{cscore.matrix}{}

\item{metadata}{}

\item{order.by.col}{}
}
\value{

}
\description{
Title
}
