% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{.plot_boxplot_samples}
\alias{.plot_boxplot_samples}
\title{(internal) plot boxplot for a given matrix of values}
\usage{
.plot_boxplot_samples(
  matrix,
  values_col_name,
  values_to_plot,
  metadata,
  color_col = NA,
  outlier
)
}
\arguments{
\item{matrix}{}

\item{values_col_name}{}

\item{values_to_plot}{}

\item{metadata}{}

\item{color_col}{}

\item{outlier}{}
}
\description{
(internal) plot boxplot for a given matrix of values
}
