% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{.plot_boxplot_sites}
\alias{.plot_boxplot_sites}
\title{Internal function for boxplot_cscore}
\usage{
.plot_boxplot_sites(matrix, values_to_plot, outlier, sort_by, horizontal)
}
\arguments{
\item{matrix}{Sites x Samples C-score/count matrix (output of extract_data()).}

\item{values_to_plot}{Value to display in plot.}

\item{outlier}{Show boxplot outlier values.}

\item{horizontal}{Show boxplot horizontally.}
}
\value{
A ggplot object.
}
\description{
Internal function for boxplot_cscore
}
\keyword{internal}
