% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{.plot_boxplot_sites}
\alias{.plot_boxplot_sites}
\title{Title}
\usage{
.plot_boxplot_sites(
  matrix,
  values_col_name,
  values_to_plot,
  metadata,
  order_by,
  outlier,
  use_renamed_positions = T
)
}
\arguments{
\item{matrix}{}

\item{values_col_name}{}

\item{values_to_plot}{}

\item{metadata}{}

\item{outlier}{}

\item{color_col}{}
}
\description{
Title
}
