% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coa.R
\name{.plot_coa}
\alias{.plot_coa}
\title{(internal) Plot a coa object}
\usage{
.plot_coa(
  dudi.coa = NULL,
  metadata = NULL,
  color_col = NULL,
  axes = c(1, 2),
  title = "default",
  subtitle = "default"
)
}
\arguments{
\item{dudi.coa}{A coa object}

\item{metadata}{RiboClass's metadata dataframe}

\item{color_col}{Name of the column in the metadata used for coloring samples.}

\item{axes}{Two-element vector indicating which pair of COA components
to show.}

\item{title}{Title to display on the plot. 'default' for default title.}

\item{subtitle}{Subtitle to display on the plot. 'samples' for number of
samples. 'none' for no subtitle.}
}
\value{
A ggplot object containing the COA
}
\description{
(internal) Plot a coa object
}
\keyword{internal}
