% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{.plot_heatmap}
\alias{.plot_heatmap}
\title{Title}
\usage{
.plot_heatmap(
  cscore_matrix = NULL,
  annotation_samples = NULL,
  cols_for_annotation = NULL,
  most_variant = F
)
}
\arguments{
\item{most_variant}{}
}
\value{

}
\description{
Title
}
