% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{.plot_heatmap}
\alias{.plot_heatmap}
\title{Internal function to plot heatmap.}
\usage{
.plot_heatmap(
  cscore_matrix = NULL,
  metadata = NULL,
  color_col = NULL,
  most_variant = F,
  only_annotated = TRUE,
  title = "default",
  cutree_rows,
  cutree_cols,
  ...
)
}
\arguments{
\item{cscore_matrix}{c-score data for all samples as a matrix (from extract_data())}

\item{metadata}{metadata for samples in cscore_matrix}

\item{color_col}{Vector of the metadata columns’ name used for coloring samples.}

\item{most_variant}{select only the most variant positions (cannot be used from plot_heatmap())}

\item{only_annotated}{Use only annotated sites (default = TRUE).}

\item{title}{Title to display on the plot. "default" for default title.}

\item{cutree_rows}{number of clusters the rows are divided into, based on the hierarchical clustering (using cutree).}

\item{cutree_cols}{number of clusters the columns are divided into, based on the hierarchical clustering (using cutree).}

\item{...}{Pheatmap’s parameters}
}
\description{
Internal function to plot heatmap.
}
\keyword{internal}
