% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_corr.R
\name{.plot_heatmap_corr}
\alias{.plot_heatmap_corr}
\title{Internal function of plot_heatmap_corr}
\usage{
.plot_heatmap_corr(cscore_matrix, metadata, color_col)
}
\arguments{
\item{cscore_matrix}{Sites x Samples C-score matrix (output of extract_data()).}

\item{metadata}{Metadata of samples in matrix}

\item{color_col}{Vector of the metadata columns’ name used for coloring samples.}
}
\value{
ComplexHeatmap heatmap
}
\description{
Internal function of plot_heatmap_corr
}
\keyword{internal}
