% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{.plot_heatmap_corr}
\alias{.plot_heatmap_corr}
\title{Internal function of plot_heatmap_corr.}
\usage{
.plot_heatmap_corr(
  cscore_matrix = NULL,
  annotation_samples = NULL,
  cols_for_annotation = NA,
  order_by_col = NULL,
  use_triangle = F
)
}
\arguments{
\item{use_triangle}{}
}
\value{

}
\description{
Internal function of plot_heatmap_corr.
}
