% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{.plot_heatmap_corr}
\alias{.plot_heatmap_corr}
\title{Internal function of plot_heatmap_corr.}
\usage{
.plot_heatmap_corr(
  cscore.matrix = NULL,
  annotation.samples = NULL,
  cols_for_annotation = NULL,
  order.by.col = NULL
)
}
\arguments{
\item{cscore.matrix}{}

\item{annotation.samples}{}

\item{cols_for_annotation}{}

\item{order.by.col}{}
}
\value{

}
\description{
Internal function of plot_heatmap_corr.
}
