% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{.plot_pca}
\alias{.plot_pca}
\title{Plot a dudi.pca object using ggplot2}
\usage{
.plot_pca(
  dudi.pca = NULL,
  metadata = NULL,
  color_col = NULL,
  axes = axes,
  title = "default",
  subtitle = "samples",
  draw_ellipses = FALSE,
  draw_centroids = FALSE,
  sample_colors = NULL
)
}
\arguments{
\item{dudi.pca}{a dudi.pca object generated with Ade4}

\item{metadata}{metadata table from SummarizedExperiment}

\item{color_col}{Name of the column in the metadata used for coloring samples.}

\item{axes}{Two-element vector indicating which pair of principal components
to show.}

\item{title}{Title to display on the plot. 'default' for default title.}

\item{subtitle}{Subtitle to display on the plot. 'samples' for number of
samples. 'none' for no subtitle.}

\item{draw_ellipses}{If TRUE, draw ellipses around groups.}

\item{draw_centroids}{If TRUE, draw group centroids on the plot.}

\item{sample_colors}{Optional named character vector mapping metadata values
in \code{color_col} to colors.}
}
\value{
a ggplot
}
\description{
Plot a dudi.pca object using ggplot2
}
\keyword{internal}
