% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{.plot_pca}
\alias{.plot_pca}
\title{Title}
\usage{
.plot_pca(dudi.pca = NULL, metadata = NULL, col.by.col = NULL)
}
\arguments{
\item{dudi.pca}{}

\item{metadata}{}

\item{col.by.col}{}
}
\value{

}
\description{
Title
}
