% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{.plot_pca}
\alias{.plot_pca}
\title{Plot a dudi.pca object using factoextra}
\usage{
.plot_pca(
  dudi.pca = NULL,
  metadata = NULL,
  color_col = NULL,
  axes = axes,
  title = NULL
)
}
\arguments{
\item{dudi.pca}{a dudi.pca object generated with Ade4}

\item{metadata}{metadata table from RiboClass}

\item{color_col}{column in the metadata used for coloring the PCA.}

\item{axes}{two-element vector indicating which pair of principal components you want to show.}
}
\value{
a ggplot
}
\description{
Plot a dudi.pca object using factoextra
}
\keyword{internal}
