% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_riboclass.R
\name{.read_count_files}
\alias{.read_count_files}
\title{Import and transform a list of count files for a RiboClass}
\usage{
.read_count_files(
  path_to_files,
  sep,
  header,
  rna_col,
  position_col,
  count_value,
  metadata_filenames = NULL
)
}
\arguments{
\item{path_to_files}{path to the folder containing count files}

\item{sep}{CSV seperator}

\item{header}{does the files contain an header ?}

\item{rna_col}{position of the column containing RNA name}

\item{position_col}{position of the column containing the genomic position}

\item{count_value}{position of the column contaning the count for the given genomic position}

\item{metadata_filenames}{Filenames in metadata to check against}
}
\value{
a list of sample dataframes
}
\description{
This internal function is used to generate the "data" part of the RiboClass.
It will read all CSV files and turn them into a list of dataframes.
}
\keyword{internal}
