% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboClass.R
\name{extract_data}
\alias{extract_data}
\title{Aggregate results into a single matrix}
\usage{
extract_data(
  ribo,
  col = "cscore",
  position_to_rownames = F,
  only_annotated = F
)
}
\arguments{
\item{ribo}{a RiboClass object}

\item{col}{column in data you want extract data from (cscore or count).}

\item{position_to_rownames}{if true, position will be included as a rowname. They will in a new column otherwise.}

\item{only_annotated}{if true, return a dataframe with only annotated sites. Return all sites otherwise.}
}
\value{
a dataframe
}
\description{
For a given column in data, this function will generate a dataframe with all samples
and all positions (if only_annotated is false) or only positions with a siteID (if only_annotated is true).
}
\examples{
data("ribo_toy")
count_df <- extract_data(ribo_toy,"count")
}
