% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_to_plot.R
\name{format_to_plot}
\alias{format_to_plot}
\title{Transfrom RiboClass data to a ggplot-friendly data frame}
\usage{
format_to_plot(ribo, metadata_col = NULL, only_annotated = FALSE)
}
\arguments{
\item{ribo}{a RiboClass object.}

\item{metadata_col}{The metadata columns to add when transforming.}

\item{only_annotated}{Keep only sites that have been annotated.}
}
\value{
A ggplot-friendly dataframe.
}
\description{
Turn a riboClass into a dataframe. You can append some metadata columns if
you need extra information for your ggplot.
}
