% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{generate_metadata_df}
\alias{generate_metadata_df}
\title{Generate a metadata dataframe, given a list of files}
\usage{
generate_metadata_df(
  counts_folder_path,
  create_samplename_col = T,
  stop_symbol = NA
)
}
\arguments{
\item{counts_folder_path:}{the path where count files are stored}

\item{create_samplename_col:}{generate a sample name col with filename by default}

\item{stop_symbol:}{keep the filename until the stop symbol is reached}
}
\description{
Generate a metadata dataframe, given a list of files
}
