% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_metadata_df.R
\name{generate_metadata_df}
\alias{generate_metadata_df}
\title{Generate a metadata dataframe, given a path.}
\usage{
generate_metadata_df(counts_folder_path, create_samplename_col = TRUE)
}
\arguments{
\item{counts_folder_path}{The path where count files are stored.}

\item{create_samplename_col}{Generate a sample name col with filename by default}
}
\value{
a data frame
}
\description{
Generate an "empty" dataframe with the required filename and samplename columns to include in a riboclass.

Names in the samplename column can be modified, as long as each sample has an unique name.

\strong{The filename column cannot be modified, unless the filenames on disk have changed.} Otherwise, it would prevent \code{\link{new_riboclass}} from linking data and metadata...
}
\keyword{internal}
