% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_annotation.R
\name{get_annotation}
\alias{get_annotation}
\title{Get annotation of a RiboClass}
\usage{
get_annotation(ribo)
}
\arguments{
\item{ribo}{A RiboClass}
}
\value{
A dataframe with the rna name, the position on the rna and the annotated site.
}
\description{
Get annotation of a RiboClass
}
\examples{
data("ribo_toy")
get_annotation(ribo_toy)
}
