% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_or_remove_samples_ribo.R
\name{keep_ribo_samples}
\alias{keep_ribo_samples}
\title{Keep only a selected number of samples}
\usage{
keep_ribo_samples(ribo = NULL, samples_to_keep = NULL)
}
\arguments{
\item{ribo}{a RiboClass object, see constructor : 
\code{\link{create_riboclass}}}

\item{samples_to_keep}{A vector with the samples to keep}
}
\value{
a riboClass with only selected samples in data and metadata
}
\description{
Keep only a selected number of samples
}
\examples{
data("ribo_toy")
ribo_test <- keep_ribo_samples(ribo = ribo_toy, samples_to_keep = c("S1", "S2"))
}
