% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ribodata.R
\name{load_ribodata}
\alias{load_ribodata}
\title{Load csv files from GenomeCov and their associated metadata. Create a RiboClass.}
\usage{
load_ribodata(
  count_path,
  metadata = NULL,
  count_sep = "\\t",
  metadata_sep = ",",
  count_header = FALSE,
  count_value = 3,
  count_rnaid = 1,
  count_pos = 2,
  metadata_key = "filename",
  metadata_id = NULL,
  flanking = 6,
  method = "median",
  ncores = 1
)
}
\arguments{
\item{count_path}{(required) path to the data folder containing count files.}

\item{metadata}{Data frame or path to a CSV file containing metadata.}

\item{count_sep}{Delimiter used for the CSV files.}

\item{metadata_sep}{Delimiter used in metadata (if imported from file).}

\item{count_header}{Boolean, specify if count files have a header or not.}

\item{count_value}{Name or index of the column containing count values.}

\item{count_rnaid}{Name or index of the column containing the name of the RNA in count data.}

\item{count_pos}{Name or index of the column containing the site's position in count data.}

\item{metadata_key}{(required when metadata not null) Name or index of the column containing the samples' filename.}

\item{metadata_id}{Name or index of the column containing the sample name.}

\item{flanking}{Size of the local coverage.}

\item{method}{Computation method of the local coverage. Either "median" or "mean".}

\item{ncores}{Number of cores to use in case of multithreading.}
}
\value{
a RiboClass
}
\description{
Import your count CSV files and the metadata to create a RiboClass.
The RiboClass is used by rRMSAnalyzer package for all analyses.

\strong{This function serves as the entrypoint of rRMSAnalyzer.}
}
\details{
load_ribodata is a wrapper of \code{\link{new_riboclass}} and \code{\link{compute_cscore}}.
}
\seealso{
new_riboclass
}
