% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ribodata.R
\name{load_ribodata}
\alias{load_ribodata}
\title{Load csv files from GenomeCov and their associated metadata. Create a RiboClass.}
\usage{
load_ribodata(
  count_path,
  metadata = NULL,
  count_sep = "\\t",
  metadata_sep = ",",
  count_header = FALSE,
  count_value = 3,
  count_rnaid = 1,
  count_pos = 2,
  metadata_key = "filename",
  metadata_id = NULL,
  flanking = 6,
  method = "median",
  ncores = 1
)
}
\arguments{
\item{count_path}{(required) path to the data folder containing count files}

\item{metadata}{data frame or path to a CSV file containing metadata.}

\item{count_sep}{delimiter used in genomecov (for csv file only)}

\item{metadata_sep}{delimiter used in metadata (for csv file only)}

\item{count_header}{boolean, specify if count files have a header or not.}

\item{count_value}{column containing count values}

\item{count_rnaid}{name or position of the column containing the name of the RNA in count data.}

\item{count_pos}{name or position of the column containing the site's position in count data.}

\item{metadata_key}{name or position of the column containing the samples' filename}

\item{metadata_id}{if it exists, name or position of the column containing the sample name}

\item{flanking}{the window size around the position (the latter is excluded)}

\item{method}{either "median" or "mean", depending how the cscore must be computed.}

\item{ncores}{number of ncores to use in case of multithreading}
}
\value{
a RiboClass
}
\description{
Import your count CSV files and the metadata to create a RiboClass.
The RiboClass is used by Riboscore package for all analyses.
}
\details{
load_ribodata is a wrapper of \code{\link{create_riboclass}} and \code{\link{compute_cscore}}.
}
