% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_samples_by_condition.R
\name{mean_samples_by_condition}
\alias{mean_samples_by_condition}
\title{Regroup samples by condition and calculate mean for each condition}
\usage{
mean_samples_by_condition(
  ribo,
  value,
  metadata_condition,
  only_annotated = FALSE
)
}
\arguments{
\item{ribo}{a RiboClass object}

\item{value}{name or index of the column containing the values on which mean
by condition is calculated. (Either "cscore" or "count")}

\item{metadata_condition}{name or index of the column \strong{in metadata} containing
the condition}

\item{only_annotated}{use annotation site name instead of default}
}
\value{
a dataframe with the mean for each condition for a selected value
}
\description{
An helper function that will give the mean by condition of the
cscore or count, for each position. The standard deviation is also given.
this can be used to create a boxplot with ggplot.
}
\examples{
data("ribo_toy")
mean_df <- mean_samples_by_condition(ribo_toy,"count","condition")
}
