% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot_count}
\alias{plot_boxplot_count}
\title{Plot a boxplot of a SummarizedExperiment counts.}
\usage{
plot_boxplot_count(ribo, color_col = NA, outlier = TRUE, horizontal = FALSE)
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{color_col}{Name of the column in the metadata used for coloring samples.}

\item{outlier}{Show boxplot outlier values.}

\item{horizontal}{Show boxplot horizontally.}
}
\value{
A ggplot object.
}
\description{
This plot is useful to check if the samples are alike in their
raw counts.
}
\examples{
data("ribo_toy")
# boxplot_count(ribo_toy,"run")

}
