% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot_count}
\alias{plot_boxplot_count}
\title{Plot a boxplot of all counts for each sample}
\usage{
plot_boxplot_count(ribo, metadata_col = NA, show_outlier = T)
}
\arguments{
\item{ribo}{a riboclass object containing your samples}

\item{metadata_col}{(optional) metadata column to colorize your boxes}
}
\value{
a ggplot object
}
\description{
This plot is useful to check if the samples are alike in their raw counts.
}
\examples{
plot_boxplot_count(ribo,"group")
}
