% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_counts_env}
\alias{plot_counts_env}
\title{Plot counts for a given position on a given RNA}
\usage{
plot_counts_env(
  ribo = NULL,
  rna = NULL,
  pos = NULL,
  samples = "all",
  flanking = 6
)
}
\arguments{
\item{ribo}{A RiboClass object}

\item{rna}{Name of RNA where the position is located}

\item{pos}{Position on RNA on which the view will be centered}

\item{samples}{Samples to display. "all" will display all samples.}

\item{flanking}{Number of sites to display on the left/right of the selected position.}
}
\description{
Plot counts for a given position on a given RNA
}
\examples{
data("ribo_toy")
ribo_toy <- rename_rna(ribo = ribo_toy)
plot_counts_env(ribo = ribo_toy, rna = "5.8S", pos = 15)
}
