% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites.R
\name{plot_diff_sites}
\alias{plot_diff_sites}
\title{plot differential sites between conditions.}
\usage{
plot_diff_sites(
  ribo,
  factor_column,
  p_cutoff = 0.01,
  cscore_cutoff = 0.05,
  statistical_test = "kruskal",
  adjust_pvalues_method = "fdr",
  sample_colors = NULL,
  object_only = FALSE
)
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{factor_column}{Metadata column used to group samples by.}

\item{p_cutoff}{Cutoff for the adjusted p-value of the kruskal-wallis test.}

\item{cscore_cutoff}{Cutoff above which the max-min c-score range between conditions' mean is considered significant.}

\item{statistical_test}{Statistical test used to compute p-values.
One of "kruskal" (wilcox is automatically used if there are 2 groups) or "t.test".}

\item{adjust_pvalues_method}{Method used to adjust p-value (one of p.adjust.methods)}

\item{sample_colors}{Optional named character vector mapping metadata values
in \code{factor_column} to colors.}

\item{object_only}{Return the results of the kruskal-wallis and C-score mean range in a dataframe directly.}
}
\value{
a ggplot object.
}
\description{
Display a boxplot between condition for each site where there is a sufficent difference between conditions.
The p-value and adjusted p-value are displayed at the bottom left of each boxplot.
}
\details{
To be considered as differential, a site must follow two conditions :
\itemize{
\item Have a significant p-value on the selected statistical
test on c-score between conditions.
\item Have a C-score range between conditions (max median - min mea)
above a certain cutoff.
}

Both the p-value cutoff and c-score range cutoff can be changed with p_cutoff and cscore_cutoff parameters respectively.
}
\examples{
data("ribo_toy")
data("human_methylated")
# ribo_toy <- rename_rna(ribo_toy)
# ribo_toy <- annotate_site(ribo_toy,human_methylated)
# plot_diff_sites(ribo_toy,"condition", p_cutoff=0.1)
}
