% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FCA.R
\name{plot_fca}
\alias{plot_fca}
\title{factorial correspondence analysis of a riboclass for a given column}
\usage{
plot_fca(ribo, col_to_plot, order_by_col, col_for_color = NULL)
}
\arguments{
\item{ribo}{A RiboClass object}

\item{col_to_plot}{column containing the values to plot the fca on}

\item{order_by_col}{column containing the sample names for linking data and metadata}

\item{col_for_color}{metadata column to colorize samples}
}
\value{
a FCA plot
}
\description{
factorial correspondence analysis of a riboclass for a given column
}
