% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_global_profile.R
\name{plot_global_profile}
\alias{plot_global_profile}
\title{Plot global profile for a given condition on a given metadata}
\usage{
plot_global_profile(ribo = ribo_adj_annot, condition_col = NULL)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{condition_col}{condition column for the plot}
}
\value{
a ggplot object
}
\description{
Plot global profile for a given condition on a given metadata
}
\examples{
plot_global_profile(ribo, condition_col, ech, ctrl)
}
