% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_global_profile.R
\name{plot_global_profile}
\alias{plot_global_profile}
\title{Plot global profile for given conditions on a given metadata}
\usage{
plot_global_profile(ribo = ribo_adj_annot, condition_col = NULL)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{condition_col}{condition column for the plot}
}
\value{
a ggplot object
}
\description{
Plot global profile for given conditions on a given metadata
}
\examples{
data("ribo_toy")
data("human_methylated")
ribo_toy <- remove_ribo_samples(ribo_toy,c("RNA1", "RNA2"))
ribo_toy <- rename_rna(ribo_toy)  
ribo_toy <- annotate_site(ribo_toy,
                                annot = human_methylated,
                                anno_rna = "rRNA",
                                anno_pos = "Position",
                                anno_value = "Nomenclature")
plot_global_profile(ribo_toy, "condition")
}
