% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot heatmap for a riboclass object}
\usage{
plot_heatmap(ribo, cols_for_annotation = NULL, most_variant = F)
}
\arguments{
\item{ribo}{a riboclass object}

\item{cols_for_annotation}{(optional) metadata columns to show on the heatmap}

\item{most_variant}{boolean to select the most variants}
}
\value{

}
\description{
This easy function will let you display an heatmap for any given column (count or c-score). You add an additionnal layer of information with metadata columns.
}
\examples{
plot_heatmap(ribo,  cols_for_annotation = c("SAMPLE..ORIGIN"), most_variant = TRUE)
}
