% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot heatmap for a ribo}
\usage{
plot_heatmap(
  ribo,
  col_to_plot,
  cols_for_annotation,
  order_by_col,
  most_variant = F
)
}
\arguments{
\item{ribo}{}

\item{col_to_plot}{}

\item{order_by_col}{}
}
\value{

}
\description{
plot heatmap for a ribo
}
