% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap for a riboclass object.}
\usage{
plot_heatmap(
  ribo,
  color_col = NULL,
  only_annotated = FALSE,
  title,
  cutree_rows = 4,
  cutree_cols = 2,
  ...
)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{color_col}{Vector of the metadata columns’ name used for coloring samples.}

\item{only_annotated}{Use only annotated sites (default = TRUE).}

\item{title}{Title to display on the plot. "default" for default title.}

\item{cutree_rows}{number of clusters the rows are divided into, based on the hierarchical clustering (using cutree).}

\item{cutree_cols}{number of clusters the columns are divided into, based on the hierarchical clustering (using cutree).}

\item{...}{Pheatmap’s parameters}
}
\value{
A ggplot object of a heatmap. The distance used is manhattan and the clustering method is Ward.D2. See pheatmap doc for more details
}
\description{
This easy function will let you display an heatmap for any given column (count or c-score). You can add an additionnal layer of information with metadata columns.
}
\examples{
data("ribo_toy")
data("human_methylated")#' 
ribo_toy <- rename_rna(ribo_toy)
ribo_toy <- annotate_site(ribo_toy,human_methylated)
plot_heatmap(ribo_toy,  color_col = c("run","condition"), only_annotated=TRUE)
}
