% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_corr.R
\name{plot_heatmap_corr}
\alias{plot_heatmap_corr}
\title{Plot a correlation heatmap from a SummarizedExperiment object.}
\usage{
plot_heatmap_corr(ribo, values_col, color_col = NULL)
}
\arguments{
\item{ribo}{A SummarizedExperiment object.}

\item{values_col}{Name of the column containing the value (either count or cscore).}

\item{color_col}{Vector of the metadata columns’ name used for coloring samples.}
}
\value{
ComplexHeatmap object
}
\description{
Shows the correlation \strong{distance} between samples.
}
\examples{
data("ribo_toy")
# plot_heatmap_corr(ribo_toy,"count","condition")

}
