% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap_corr}
\alias{plot_heatmap_corr}
\title{plot a correlation heatmap from a riboclass object}
\usage{
plot_heatmap_corr(
  ribo,
  col_to_plot,
  cols_for_annotation,
  order_by_col,
  use_triangle = F
)
}
\arguments{
\item{ribo}{}

\item{col_to_plot}{}

\item{cols_for_annotation}{}

\item{order_by_col}{}
}
\value{

}
\description{
plot a correlation heatmap from a riboclass object
}
