% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap_corr}
\alias{plot_heatmap_corr}
\title{plot a correlation heatmap from a riboclass object}
\usage{
plot_heatmap_corr(
  ribo,
  col_to_plot,
  cols_for_annotation,
  order_by_col,
  use_triangle = F
)
}
\arguments{
\item{ribo}{a riboclass object}

\item{col_to_plot}{column containing the values to be used for the heatmap.}

\item{cols_for_annotation}{metadata columns for annotation}

\item{order_by_col}{column that contains your sample names. Makes the link between your data and metadata}

\item{use_triangle}{Show only a correlation triange without any more values}
}
\value{

}
\description{
This heatmap is used to compare
}
