% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_corr.R
\name{plot_heatmap_corr}
\alias{plot_heatmap_corr}
\title{Plot a correlation heatmap from a riboclass object.

Shows the correlation \strong{distance} between samples.}
\usage{
plot_heatmap_corr(ribo, values_col, color_col)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{color_col}{Vector of the metadata columns’ name used for coloring samples.}
}
\value{
ComplexHeatmap object
}
\description{
Plot a correlation heatmap from a riboclass object.

Shows the correlation \strong{distance} between samples.
}
\examples{
data("ribo_toy")
plot_heatmap_corr(ribo_toy,"count","condition")

}
