% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites.R
\name{plot_most_differential_sites}
\alias{plot_most_differential_sites}
\title{(Internal) Display differential sites.}
\usage{
plot_most_differential_sites(
  df_of_Cscores = NULL,
  df_of_kruskal = NULL,
  most_differential_sites = NULL
)
}
\arguments{
\item{df_of_Cscores}{Dataframe of positions x samples for c-score. Output of extract_data().}

\item{df_of_kruskal}{Output of kruskal_test_on_cscores()}

\item{most_differential_sites}{}
}
\description{
Using dataframes obtained from
}
\keyword{internal}
