% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Principal component analysis of a RiboClass object}
\usage{
plot_PCA(
  ribo,
  color_col = NULL,
  axes = c(1, 2),
  only_annotated = F,
  pca_object_only = F
)
}
\arguments{
\item{ribo}{a RiboClass object}

\item{color_col}{column in the metadata used for coloring the PCA.}

\item{axes}{two-element vector indicating which pair of principal components you want to show.}

\item{only_annotated}{Use only annotated sites to plot PCA.}

\item{pca_object_only}{Return directly the full dudi.pca object, without generating the plot.}
}
\value{
a ggplot or a dudi.pca object if pca_object_only is set to True.
}
\description{
Principal component analysis of a RiboClass object
}
\examples{
data("ribo_toy")
plot_PCA(ribo_toy,"run")
}
