% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Principal component analysis of a RiboClass object}
\usage{
plot_pca(
  ribo,
  color_col = NULL,
  axes = c(1, 2),
  only_annotated = FALSE,
  title = "default",
  subtitle = "samples",
  object_only = FALSE
)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{color_col}{Name of the column in the metadata used for coloring samples.}

\item{axes}{Two-element vector indicating which pair of principal components you want to show.}

\item{only_annotated}{If TRUE, use only annotated sites to plot PCA.}

\item{title}{Title to display on the plot. "default" for default title.}

\item{subtitle}{Subtitle to display on the plot. "samples" for number of samples. "none" for no subtitle.}

\item{object_only}{Return directly the full dudi.pca object, without generating the plot.}
}
\value{
A ggplot or a dudi.pca object if object_only is set to True.
}
\description{
Principal component analysis of a RiboClass object
}
\examples{
data("ribo_toy")
plot_pca(ribo_toy,"run")
}
