% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Title}
\usage{
plot_PCA(ribo, col_to_plot, order_by_col, col_for_color = NULL)
}
\arguments{
\item{ribo}{}

\item{col_to_plot}{}

\item{order_by_col}{}

\item{col_for_color}{}
}
\value{

}
\description{
Title
}
