% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rle.R
\name{plot_rle}
\alias{plot_rle}
\title{Relative Log Expression plot of a RiboClass object’s counts.}
\usage{
plot_rle(ribo, show_outlier = TRUE)
}
\arguments{
\item{ribo}{a RiboClass object.}

\item{outlier}{Show boxplot outlier values.}
}
\value{
A ggplot object. Samples with a median lower than median(RLE counts)-2*MAD (Median Absolute Deviation) are colored in red.
}
\description{
Relative Log Expression plot of a RiboClass object’s counts.
}
\examples{
data("ribo_toy")
plot_rle(ribo_toy)
}
