% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat.R
\name{plot_stat}
\alias{plot_stat}
\title{Plot statistic of significant differential sites between experimental condition returned by the 
res.pv function}
\usage{
plot_stat(
  ribo = ribo,
  p_cutoff = 0.05,
  cscore_cutoff = 0.5,
  adjust_pvalues_method = "fdr",
  site = NULL,
  res_pv = res_pv,
  pthr = NULL,
  condition_col = NULL
)
}
\arguments{
\item{ribo}{a RiboClass}

\item{p_cutoff}{cutoff of p-value under which sites are not taken in account}

\item{cscore_cutoff}{cutoff of c-score under which sites are not taken in account}

\item{adjust_pvalues_method}{Method to adjust p-value}

\item{site}{sites to plot}

\item{res_pv}{df of p values extracted from res_pv.R}

\item{pthr}{p value threshold}

\item{condition_col}{the condition column passed in params of the Rmd}
}
\value{
a ggplot object
}
\description{
Plot statistic of significant differential sites between experimental condition returned by the 
res.pv function
}
\examples{
data("ribo_toy")
data("human_methylated")
ribo_toy <- remove_ribo_samples(ribo_toy,c("RNA1", "RNA2"))
ribo_toy <- rename_rna(ribo_toy)  
ribo_toy <- annotate_site(ribo_toy,
                                annot = human_methylated,
                                anno_rna = "rRNA",
                                anno_pos = "Position",
                                anno_value = "Nomenclature")
res_pv <- res_pv(ribo = ribo_toy, test = "kruskal", condition_col = "condition") 
plot_stat(ribo = ribo_toy, site = NULL, res_pv = res_pv, pthr = 0.01, 
condition_col = "condition", cscore_cutoff = 0.5)
}
