% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_site.R
\name{remove_annotation}
\alias{remove_annotation}
\title{Remove site annotations of a given RiboClass}
\usage{
remove_annotation(ribo)
}
\arguments{
\item{ribo}{A RiboClass object.}
}
\value{
A RiboClass object where all annotated position have been replaced by
NA, the default value.
}
\description{
Remove site annotations of a given RiboClass
}
\examples{
data('ribo_toy')
remove_annotation(ribo_toy)
}
