% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_or_remove_samples_ribo.R
\name{remove_ribo_samples}
\alias{remove_ribo_samples}
\title{Remove samples from Riboclass object}
\usage{
remove_ribo_samples(ribo = NULL, samples_to_delete = NULL)
}
\arguments{
\item{ribo}{a RiboClass}

\item{samples_to_delete}{A vector with the samples to remove}
}
\value{
a riboClass without the removed samples in data and metadata
}
\description{
Remove samples from Riboclass object
}
\examples{
data("ribo_toy")
ribo.test <- remove_ribo_samples(ribo = ribo_toy, samples_to_delete = c("4283", "3307"))
}
