% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rna.R
\name{remove_rna}
\alias{remove_rna}
\title{Remove one or more RNA among your RiboClass' samples}
\usage{
remove_rna(ribo, rna_to_remove)
}
\arguments{
\item{ribo}{A RiboClass object.}

\item{rna_to_remove}{One RNA name or a vector of RNA names.}
}
\value{
A RiboClass with the RNA(s) removed in data and rna_names sections.
}
\description{
Remove one or more RNA among your RiboClass' samples
}
\examples{
data("ribo_toy")
ribo_toy <- remove_rna(ribo_toy, "NR_023363.1_5S")
}
