% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_rna.R
\name{rename_rna}
\alias{rename_rna}
\title{Rename RNAs in your RiboClass}
\usage{
rename_rna(ribo, new_names = c("5S", "5.8S", "18S", "28S"))
}
\arguments{
\item{ribo}{a RiboClass object}

\item{new_names}{the new names for your RNA (by order of rna size)}
}
\value{
a RiboClass with updated rna names in data.
}
\description{
Rename RNAs in your RiboClass
}
\examples{
ribo_toy <- rename_rna(ribo_toy ,c("5S","5.8S","18S","28S"))
}
