% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_check.R
\name{report_qc}
\alias{report_qc}
\title{Generate html report about batch effect in the working directory}
\usage{
report_qc(ribo, library_col, output_dir = getwd())
}
\arguments{
\item{ribo}{a riboclass}

\item{library_col}{library column in the metadata}
}
\description{
Generate html report about batch effect in the working directory
}
