% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_qc}
\alias{report_qc}
\title{Generate html report about possible batch effect for a given RiboClass.}
\usage{
report_qc(
  ribo,
  specie = "not specified",
  library_col,
  project_name = "Unnamed project",
  output_dir = getwd(),
  comments = NULL
)
}
\arguments{
\item{ribo}{a RiboClass object.}

\item{specie}{studied specie of the project}

\item{library_col}{name or index of the column __in metadata__ containing the library}

\item{project_name}{Name of the project to display on the report}

\item{output_dir}{Path of output directory}

\item{comments}{Path to a text file containing comments}
}
\value{
an html file
}
\description{
Generate html report about possible batch effect for a given RiboClass.
}
\examples{
data('ribo_toy')
report_qc(ribo_toy, "human", "run")
}
