% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_qc.R
\name{report_qc}
\alias{report_qc}
\title{Generate html report about batch effect in the working directory}
\usage{
report_qc(ribo, library_col, project_name = "", output_dir = getwd())
}
\arguments{
\item{ribo}{a riboclass}

\item{library_col}{library column in the metadata}

\item{project_name}{Name of the project}
}
\description{
Generate html report about batch effect in the working directory
}
