% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboClass.R
\name{ribo_remove_rna}
\alias{ribo_remove_rna}
\title{Remove a RNA among all samples}
\usage{
ribo_remove_rna(ribo, name_rna_to_remove)
}
\arguments{
\item{ribo}{a RiboClass}

\item{name_rna_to_remove}{name of the rna to remove}
}
\value{
a RiboClass without the specified RNA
}
\description{
Remove a RNA among all samples
}
