% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ribo_toy}
\alias{ribo_toy}
\title{RiboClass from a toy dataset}
\format{
a RiboClass (S3 Class) with the following element
\describe{
  \item{data}{list of sample dataframe}
  \item{metadata}{metadata dataframe of all samples}
  \item{rna_names}{dataframe containing both original and modified RNAs' name}
  }
}
\usage{
ribo_toy
}
\description{
A riboclass containing 10 samples + 2 reference RNA.
}
\details{
Samples are from 4 different biological conditions ("condition" column in metadata). The sequencing has been 
done in two different batches ("run" column in metadata).
Both batches have the same reference RNA, to detect technical bias.
}
\keyword{datasets}
