% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites.R
\name{select_most_differential_sites}
\alias{select_most_differential_sites}
\title{Return a list of differential sites.}
\usage{
select_most_differential_sites(
  df_of_kruskal = NULL,
  p_cutoff = 0.01,
  cscore_cutoff = 0.05
)
}
\arguments{
\item{df_of_kruskal}{Output of kruskal_test_on_cscores().}

\item{p_cutoff}{Cutoff below which the kruskal-wallis test is considered
significant for a given site.}

\item{cscore_cutoff}{Cutoff above which the max-min c-score range between
conditions' mean is considered significant.}
}
\value{
a vector of sites that are differentials
}
\description{
Return a list of differential sites.
}
\details{
To be considered as differential, a site must follow two conditions :
\itemize{
\item Have a significant p-value on a Kruskal-wallis
test on c-score between conditions.
\item Have a C-score range between conditions (max median - min mea)
above a certain cutoff.
}

Both the p-value cutoff and c-score range cutoff can be changed with p_cutoff and cscore_cutoff parameters respectively.
}
\keyword{internal}
