% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rna_names.R
\name{show_RNA_names}
\alias{show_RNA_names}
\title{Display RNA names}
\usage{
show_RNA_names(ribo = NULL)
}
\arguments{
\item{ribo}{a RiboClass object}
}
\value{
A vector with actual RNA names
}
\description{
Display RNA names
}
\examples{
show_RNA_names(ribo = ribo_toy)
}
