% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_test.R
\name{wrapper_kruskal_test}
\alias{wrapper_kruskal_test}
\title{Test which sites are the most variable between conditions}
\usage{
wrapper_kruskal_test(
  ribo = NULL,
  order_by_col = NULL,
  adjust_pvalues_method = "fdr",
  factor_column = NULL
)
}
\arguments{
\item{ribo}{a RiboClass}

\item{adjust_pvalues_method}{p-value adjustment method (default : "fdr")}

\item{factor_column}{which column to use for factoring}
}
\description{
This function use different tests
}
